core_crypto
Module e2e_identity

re-export rusty-jwt-tools API

Modules

enabled

Utility for clients to get the current state of E2EI when the app resumes

types

We only expose byte arrays through the FFI so we do all the conversions here

Structs

CrlRegistration

Supporting struct for CRL registration result

E2eiDumpedPkiEnv

Dump of the PKI environemnt as PEM

E2eiEnrollment

Wire end to end identity solution for fetching a x509 certificate which identifies a client.

Enums

Error

Type Aliases

Result