Enum core_crypto::e2e_identity::error::E2eIdentityError

pub enum E2eIdentityError {

    ImplementationError,
    NotYetSupported,
    MissingExistingClient(MlsCredentialType),
    InvalidIdentity,
    InvalidSignatureKey,
    OutOfOrderEnrollment(&'static str),
    E2eiInvalidDomain,
    InvalidRefreshToken,
    KeyStoreError(CryptoKeystoreError),
    CryptoError(CryptoError),
    IdentityError(E2eIdentityError),
    X509Error(RustyX509CheckError),
    UrlError(ParseError),
    JsonError(Error),
    Utf8Error(Utf8Error),
    LockPoisonError,
    TrustAnchorAlreadyRegistered,

}

End to end identity errors

Variants

ImplementationError

Client misused this library

NotYetSupported

Incoming support

MissingExistingClient(MlsCredentialType)

The required local MLS client was not initialized. It’s likely a consumer error

InvalidIdentity

Cannot read the identity in the EE certificate

InvalidSignatureKey

Failed converting the MLS signature key for the e2ei enrollment

OutOfOrderEnrollment(&'static str)

Enrollment methods are called out of order

E2eiInvalidDomain

Error when an end-to-end-identity domain is not well-formed utf-16, which means it’s out of spec

InvalidRefreshToken

Invalid OIDC RefreshToken supplied

KeyStoreError(CryptoKeystoreError)

An error occurred while trying to persist the RefreshToken in the keystore

CryptoError(CryptoError)

Error generating keys

IdentityError(E2eIdentityError)

Error creating client Dpop token or acme error

X509Error(RustyX509CheckError)

Error validating X509 parameters

UrlError(ParseError)

Error parsing a URL

JsonError(Error)

Json error

Utf8Error(Utf8Error)

Utf8 error

LockPoisonError

!!!! Something went very wrong and one of our locks has been poisoned by an in-thread panic !!!!

TrustAnchorAlreadyRegistered

We already have an ACME Root Trust Anchor registered. Cannot proceed but this is usually indicative of double registration and can be ignored