core_crypto::e2e_identity

Module pki_env_hooks

PKI Environment API Hooks

Structs

HttpHeader

An http header

HttpResponse

An HTTP Response

PkiEnvironmentHooksError

Error type for PKI environment hooks

Enums

HttpMethod

An http method

Traits

PkiEnvironmentHooks

The PKI Environment Hooks used for external calls during e2e enrollment flow. When communicating with the Identity Provider (IDP) and Wire server, CoreCrypto delegates to the client app by calling the relevant methods.