Struct core_crypto::e2e_identity::E2eiEnrollment

pub struct E2eiEnrollment { /* private fields */ }

Wire end to end identity solution for fetching a x509 certificate which identifies a client.

Implementations

impl E2eiEnrollment

pub fn get_refresh_token(&self) -> E2eIdentityResult<&str>

Lets clients retrieve the OIDC refresh token to try to renew the user’s authorization. If it’s expired, the user needs to reauthenticate and they will update the refresh token in E2eiEnrollment::new_oidc_challenge_request

impl E2eiEnrollment

pub fn try_new( client_id: ClientId, display_name: String, handle: String, team: Option<String>, expiry_sec: u32, backend: &MlsCryptoProvider, ciphersuite: MlsCiphersuite, sign_keypair: Option<E2eiSignatureKeypair>, refresh_token: Option<RefreshToken> ) -> CryptoResult<Self>

Builds an instance holding private key material. This instance has to be used in the whole enrollment process then dropped to clear secret key material.

Parameters

client_id - client identifier e.g. b7ac11a4-8f01-4527-af88-1c30885a7931:6add501bacd1d90e@example.com

display_name - human readable name displayed in the application e.g. Smith, Alice M (QA)

handle - user handle e.g. alice.smith.qa@example.com

expiry_sec - generated x509 certificate expiry in seconds

pub fn directory_response( &mut self, directory: Vec<u8> ) -> E2eIdentityResult<E2eiAcmeDirectory>

Parses the response from GET /acme/{provisioner-name}/directory. Use this types::E2eiAcmeDirectory in the next step to fetch the first nonce from the acme server. Use [types::E2eiAcmeDirectory.new_nonce].

See RFC 8555 Section 7.1.1

Parameters

directory - http response body

pub fn new_account_request( &self, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

For creating a new acme account. This returns a signed JWS-alike request body to send to POST /acme/{provisioner-name}/new-account.

See RFC 8555 Section 7.3.

Parameters

directory - you got from Self::directory_response

previous_nonce - you got from calling HEAD {directory.new_nonce}

pub fn new_account_response( &mut self, account: Vec<u8> ) -> E2eIdentityResult<()>

Parses the response from POST /acme/{provisioner-name}/new-account.

See RFC 8555 Section 7.3.

Parameters

account - http response body

pub fn new_order_request( &self, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Creates a new acme order for the handle (userId + display name) and the clientId.

See RFC 8555 Section 7.4.

Parameters

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/new-account

pub fn new_order_response( &self, order: Vec<u8> ) -> E2eIdentityResult<E2eiNewAcmeOrder>

Parses the response from POST /acme/{provisioner-name}/new-order.

See RFC 8555 Section 7.4.

Parameters

new_order - http response body

pub fn new_authz_request( &self, url: String, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Creates a new authorization request.

See RFC 8555 Section 7.5.

Parameters

url - one of the URL in new order’s authorizations (from Self::new_order_response)

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/new-order (or from the previous to this method if you are creating the second authorization)

pub fn new_authz_response( &mut self, authz: Vec<u8> ) -> E2eIdentityResult<E2eiNewAcmeAuthz>

Parses the response from POST /acme/{provisioner-name}/authz/{authz-id}

See RFC 8555 Section 7.5.

Parameters

new_authz - http response body

pub fn create_dpop_token( &self, expiry_secs: u32, backend_nonce: String ) -> E2eIdentityResult<String>

Generates a new client Dpop JWT token. It demonstrates proof of possession of the nonces (from wire-server & acme server) and will be verified by the acme server when verifying the challenge (in order to deliver a certificate).

Then send it to POST /clients/{id}/access-token on wire-server.

Parameters

expiry_secs - of the client Dpop JWT. This should be equal to the grace period set in Team Management

backend_nonce - you get by calling GET /clients/token/nonce on wire-server. See endpoint definition

expiry - token expiry

pub fn new_dpop_challenge_request( &self, access_token: String, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Creates a new challenge request.

See RFC 8555 Section 7.5.1.

Parameters

access_token - returned by wire-server from this endpoint

dpop_challenge - you found after Self::new_authz_response

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/authz/{authz-id}

pub fn new_dpop_challenge_response( &self, challenge: Vec<u8> ) -> E2eIdentityResult<()>

Parses the response from POST /acme/{provisioner-name}/challenge/{challenge-id} for the DPoP challenge

See RFC 8555 Section 7.5.1.

Parameters

challenge - http response body

pub fn new_oidc_challenge_request( &mut self, id_token: String, refresh_token: String, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Creates a new challenge request.

See RFC 8555 Section 7.5.1.

Parameters

id_token - you get back from Identity Provider

refresh_token - you get back from Identity Provider to renew the access token

oidc_challenge - you found after Self::new_authz_response

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/authz/{authz-id}

pub async fn new_oidc_challenge_response( &mut self, backend: &MlsCryptoProvider, challenge: Vec<u8> ) -> E2eIdentityResult<()>

Parses the response from POST /acme/{provisioner-name}/challenge/{challenge-id} for the OIDC challenge

See RFC 8555 Section 7.5.1.

Parameters

challenge - http response body

pub fn check_order_request( &self, order_url: String, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Verifies that the previous challenge has been completed.

See RFC 8555 Section 7.4.

Parameters

order_url - location header from http response you got from Self::new_order_response

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/challenge/{challenge-id}

pub fn check_order_response( &mut self, order: Vec<u8> ) -> E2eIdentityResult<String>

Parses the response from POST /acme/{provisioner-name}/order/{order-id}.

See RFC 8555 Section 7.4.

Parameters

order - http response body

Returns

The finalize url to use with Self::finalize_request

pub fn finalize_request( &mut self, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Final step before fetching the certificate.

See RFC 8555 Section 7.4.

Parameters

domains - you want to generate a certificate for e.g. ["wire.com"]

order - you got from Self::check_order_response

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/order/{order-id}

pub fn finalize_response( &mut self, finalize: Vec<u8> ) -> E2eIdentityResult<String>

Parses the response from POST /acme/{provisioner-name}/order/{order-id}/finalize.

See RFC 8555 Section 7.4.

Parameters

finalize - http response body

Returns

The certificate url to use with Self::certificate_request

pub fn certificate_request( &mut self, previous_nonce: String ) -> E2eIdentityResult<Vec<u8>>

Creates a request for finally fetching the x509 certificate.

See RFC 8555 Section 7.4.2.

Parameters

finalize - you got from Self::finalize_response

account - you got from Self::new_account_response

previous_nonce - replay-nonce response header from POST /acme/{provisioner-name}/order/{order-id}/finalize