Enum core_crypto::CryptoError

pub enum CryptoError {

    ConversationNotFound(ConversationId),
    ClientNotFound(ClientId),
    PendingProposalNotFound(MlsProposalRef),
    PendingCommitNotFound,
    MalformedIdentifier(String),
    ClientSignatureNotFound,
    ClientSignatureMismatch,
    LockPoisonError,
    ImplementationError,
    OutOfKeyPackage(MemberId),
    MlsProviderError(MlsProviderError),
    KeyStoreError(CryptoKeystoreError),
    MlsError(MlsError),
    Utf8Error(Utf8Error),
    StringUtf8Error(FromUtf8Error),
    ParseIntError(ParseIntError),
    ConvertIntError(TryFromIntError),
    HexDecodeError(FromHexError),
    InvalidByteArrayError(usize),
    IoError(Error),
    Unauthorized,
    CallbacksNotSet,
    UnauthorizedExternalAddProposal,
    UnauthorizedExternalCommit,
    InvalidHashReference,
    GenerationOutOfBound,
    DecryptionError,
    WrongEpoch,
    ProteusError(ProteusError),
    CryptoboxMigrationError(CryptoboxMigrationError),
    ProteusNotInitialized,
    ProteusSupportNotEnabled(String),
    MlsNotInitialized,
    InvalidKeyPackage,

}

CoreCrypto errors

Variants

ConversationNotFound(ConversationId)

This error is emitted when the requested conversation couldn’t be found in our store

ClientNotFound(ClientId)

This error is emitted when the requested client couldn’t be found in MLS group

PendingProposalNotFound(MlsProposalRef)

This error is emitted when a pending proposal couldn’t be found in MLS group

PendingCommitNotFound

This error is emitted when a pending commmit couldn’t be found in MLS group

MalformedIdentifier(String)

This error is emitted when we find a malformed (i.e. not uuid) or empty identifier

ClientSignatureNotFound

The keystore has no knowledge of such client; this shouldn’t happen as Client::init is failsafe (find-else-create)

ClientSignatureMismatch

The keystore has found the client, but the provided signature doesn’t match against what is stored

LockPoisonError

!!!! Something went very wrong and one of our locks has been poisoned by an in-thread panic !!!!

ImplementationError

We have done something terribly wrong

OutOfKeyPackage(MemberId)

A conversation member is out of local stored keypackages - if it does happen something went wrong

MlsProviderError(MlsProviderError)

Errors that are sent by our MLS Provider

KeyStoreError(CryptoKeystoreError)

Errors that are sent by our Keystore

MlsError(MlsError)

MLS Internal Errors

Utf8Error(Utf8Error)

Error when parsing strs that are not UTF-8

StringUtf8Error(FromUtf8Error)

Error when parsing Strings that are not UTF-8

ParseIntError(ParseIntError)

Error when trying to coerce ints into Strings

ConvertIntError(TryFromIntError)

Error when trying to convert integer sizes - usually when they don’t fit

HexDecodeError(FromHexError)

Error when trying to decode an hex-encoded string. Usually that means that the length of the hex string is odd - should be always even

InvalidByteArrayError(usize)

Error when trying to coerce a Vec into a [u8 ; N]

IoError(Error)

Standard I/O Error

Unauthorized

Authorization error

CallbacksNotSet

Callbacks are not informed

UnauthorizedExternalAddProposal

External Add Proposal Validation failed

UnauthorizedExternalCommit

External Commit sender was not authorized to perform such

InvalidHashReference

A supplied [HashReference] is not of the expected size: 16

GenerationOutOfBound

The client who created this message is likely to have reused a ratchet generation or we are trying to decrypt the same message twice. For the former, this client MLS implementation has a loophole and it’d better be evicted from the group.

DecryptionError

Tried to decrypt a message in the wrong epoch

WrongEpoch

Incoming message is for the wrong epoch

ProteusError(ProteusError)

Proteus Error Wrapper

CryptoboxMigrationError(CryptoboxMigrationError)

Cryptobox migration error wrapper

ProteusNotInitialized

The proteus client has been called but has not been initialized yet

ProteusSupportNotEnabled(String)

CoreCrypto hasn’t been built with the proteus feature enabled, meaning proteus isn’t built in

MlsNotInitialized

A MLS operation was requested but MLS hasn’t been initialized on this instance

InvalidKeyPackage

Decrypted message uses an invalid KeyPackage (probably expired)