Enum core_crypto::MlsError

pub enum MlsError {

    MlsWelcomeError(WelcomeError),
    MlsLibraryError(LibraryError),
    MlsInvalidMessageError(CreateMessageError),
    MlsEmptyInputError(EmptyInputError),
    MlsCredentialError(CredentialError),
    MlsNewGroupError(NewGroupError),
    MlsAddMembersError(AddMembersError),
    MlsRemoveMembersError(RemoveMembersError),
    MlsUnverifiedMessageError(UnverifiedMessageError),
    MlsParseMessageError(ParseMessageError),
    MlsKeyPackageBundleNewError(KeyPackageBundleNewError),
    MlsSelfUpdateError(SelfUpdateError),
    MlsMlsGroupStateError(MlsGroupStateError),
    MlsMessageError(MlsMessageError),
    ProposeAddMemberError(ProposeAddMemberError),
    ProposeSelfUpdateError(ProposeSelfUpdateError),
    ProposeRemoveMemberError(ProposeRemoveMemberError),
    MlsCommitToPendingProposalsError(CommitToPendingProposalsError),
    MlsExportPublicGroupStateError(ExportPublicGroupStateError),
    MlsTlsCodecError(Error),
    MlsKeystoreSerializationError(Error),
    MlsErrorString(ErrorString),
    MlsExternalCommitError(ExternalCommitError),
    MlsCryptoError(CryptoError),
    MlsExportSecretError(ExportSecretError),

}

MLS-specific error wrapper - see github.com/openmls/openmls for details

Variants

MlsWelcomeError(WelcomeError)

Welcome error

MlsLibraryError(LibraryError)

Generic error type that indicates unrecoverable errors in the library. See [openmls::error::LibraryError]

MlsInvalidMessageError(CreateMessageError)

Create message error

MlsEmptyInputError(EmptyInputError)

EmptyInput error

MlsCredentialError(CredentialError)

An error that occurs in methods of a openmls::credentials::Credential.

MlsNewGroupError(NewGroupError)

New group error

MlsAddMembersError(AddMembersError)

Add members error

MlsRemoveMembersError(RemoveMembersError)

Remove members error

MlsUnverifiedMessageError(UnverifiedMessageError)

Unverified message error

MlsParseMessageError(ParseMessageError)

Parse message error

MlsKeyPackageBundleNewError(KeyPackageBundleNewError)

[openmls::key_packages::KeyPackageBundle] new error

MlsSelfUpdateError(SelfUpdateError)

Self update error

MlsMlsGroupStateError(MlsGroupStateError)

Group state error

MlsMessageError(MlsMessageError)

MlsMessage error

ProposeAddMemberError(ProposeAddMemberError)

Propose add members error

ProposeSelfUpdateError(ProposeSelfUpdateError)

Propose self update error

ProposeRemoveMemberError(ProposeRemoveMemberError)

Propose remove members error

MlsCommitToPendingProposalsError(CommitToPendingProposalsError)

Commit to pending proposals error

MlsExportPublicGroupStateError(ExportPublicGroupStateError)

Export public group state error

MlsTlsCodecError(Error)

Errors that are thrown by TLS serialization crate.

MlsKeystoreSerializationError(Error)

This type represents all possible errors that can occur when serializing or deserializing JSON data.

MlsErrorString(ErrorString)

A wrapper struct for an error string. This can be used when no complex error variant is needed.

MlsExternalCommitError(ExternalCommitError)

External Commit error

MlsCryptoError(CryptoError)

OpenMls crypto error

MlsExportSecretError(ExportSecretError)

OpenMls Export Secret error