core_crypto
Enum MlsErrorKind

pub enum MlsErrorKind {

    MlsWelcomeError(WelcomeError<CryptoKeystoreError>),
    MlsLibraryError(LibraryError),
    MlsInvalidMessageError(CreateMessageError),
    MlsEmptyInputError(EmptyInputError),
    MlsCredentialError(CredentialError),
    MlsNewGroupError(NewGroupError<CryptoKeystoreError>),
    MlsAddMembersError(AddMembersError<CryptoKeystoreError>),
    MlsRemoveMembersError(RemoveMembersError<CryptoKeystoreError>),
    MlsMessageError(ProcessMessageError),
    MlsKeyPackageBundleNewError(KeyPackageNewError<CryptoKeystoreError>),
    MlsSelfUpdateError(SelfUpdateError<CryptoKeystoreError>),
    MlsMlsGroupStateError(MlsGroupStateError),
    ProposeAddMemberError(ProposeAddMemberError),
    ProposeSelfUpdateError(ProposeSelfUpdateError<CryptoKeystoreError>),
    ProposeRemoveMemberError(ProposeRemoveMemberError),
    MlsCommitToPendingProposalsError(CommitToPendingProposalsError<CryptoKeystoreError>),
    MlsExportGroupInfoError(ExportGroupInfoError),
    MlsTlsCodecError(Error),
    MlsKeystoreSerializationError(Error),
    MlsErrorString(ErrorString),
    MlsExternalCommitError(ExternalCommitError),
    MlsCryptoError(CryptoError),
    MlsExportSecretError(ExportSecretError),
    MlsMergeCommitError(MergeCommitError<CryptoKeystoreError>),
    MlsKeyPackageValidationError(KeyPackageVerifyError),
    MlsMergePendingCommitError(MergePendingCommitError<CryptoKeystoreError>),
    MlsEncryptMessageError(MlsMessageError),
    MlsDeleteKeyPackageError(KeyPackageDeleteError<CryptoKeystoreError>),
    MlsUpdateExtensionsError(UpdateExtensionsError<CryptoKeystoreError>),
    MlsLeafNodeValidationError(LeafNodeValidationError),
    RatchetTreeError(RatchetTreeError),
    GroupInfoError(GroupInfoError),
    ProviderError(MlsProviderError),

}

Openmls produces these kinds of error

Variants

MlsWelcomeError(WelcomeError<CryptoKeystoreError>)

Welcome error

MlsLibraryError(LibraryError)

Generic error type that indicates unrecoverable errors in the library. See [openmls::error::LibraryError]

MlsInvalidMessageError(CreateMessageError)

Create message error

MlsEmptyInputError(EmptyInputError)

EmptyInput error

MlsCredentialError(CredentialError)

An error that occurs in methods of a openmls::credentials::Credential.

MlsNewGroupError(NewGroupError<CryptoKeystoreError>)

New group error

MlsAddMembersError(AddMembersError<CryptoKeystoreError>)

Add members error

MlsRemoveMembersError(RemoveMembersError<CryptoKeystoreError>)

Remove members error

MlsMessageError(ProcessMessageError)

Parse message error

MlsKeyPackageBundleNewError(KeyPackageNewError<CryptoKeystoreError>)

KeyPackageBundle new error

MlsSelfUpdateError(SelfUpdateError<CryptoKeystoreError>)

Self update error

MlsMlsGroupStateError(MlsGroupStateError)

Group state error

ProposeAddMemberError(ProposeAddMemberError)

Propose add members error

ProposeSelfUpdateError(ProposeSelfUpdateError<CryptoKeystoreError>)

Propose self update error

ProposeRemoveMemberError(ProposeRemoveMemberError)

Propose remove members error

MlsCommitToPendingProposalsError(CommitToPendingProposalsError<CryptoKeystoreError>)

Commit to pending proposals error

MlsExportGroupInfoError(ExportGroupInfoError)

Export public group state error

MlsTlsCodecError(Error)

Errors that are thrown by TLS serialization crate.

MlsKeystoreSerializationError(Error)

This type represents all possible errors that can occur when serializing or deserializing JSON data.

MlsErrorString(ErrorString)

A wrapper struct for an error string. This can be used when no complex error variant is needed.

MlsExternalCommitError(ExternalCommitError)

External Commit error

MlsCryptoError(CryptoError)

OpenMls crypto error

MlsExportSecretError(ExportSecretError)

OpenMls Export Secret error

MlsMergeCommitError(MergeCommitError<CryptoKeystoreError>)

OpenMLS merge commit error

MlsKeyPackageValidationError(KeyPackageVerifyError)

OpenMLS keypackage validation error

MlsMergePendingCommitError(MergePendingCommitError<CryptoKeystoreError>)

OpenMLS Commit merge error

MlsEncryptMessageError(MlsMessageError)

OpenMLS encrypt message error

MlsDeleteKeyPackageError(KeyPackageDeleteError<CryptoKeystoreError>)

OpenMLS delete KeyPackage error

MlsUpdateExtensionsError(UpdateExtensionsError<CryptoKeystoreError>)

OpenMLS update extensions error

MlsLeafNodeValidationError(LeafNodeValidationError)

OpenMLS LeafNode validation error

RatchetTreeError(RatchetTreeError)

OpenMLS LeafNode validation error

GroupInfoError(GroupInfoError)

OpenMLS GroupInfo error

ProviderError(MlsProviderError)

Provider Error