Trait core_crypto::keypairs::Keypair

pub trait Keypair: ConstantTimeEq + ZeroizeOnDrop + Sized {
    type SecretLen: ArrayLength<u8>;
    type Public: BinarySerializable + Clone + PartialEq;

    // Required methods
    fn random() -> Self;
    fn from_secret(bytes: &[u8]) -> Result<Self>;
    fn secret(&self) -> &SecretValue<Self::SecretLen>;
    fn public(&self) -> &Self::Public;

    // Provided method
    fn unzip(self) -> (SecretValue<Self::SecretLen>, Self::Public) { ... }
}

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

Required Associated Types

type SecretLen: ArrayLength<u8>

Represents the type of the private (secret) key

type Public: BinarySerializable + Clone + PartialEq

Represents the type of the public key

Required Methods

fn random() -> Self

Generates a new random keypair.

fn from_secret(bytes: &[u8]) -> Result<Self>

Creates a keypair from the given secret key.

fn secret(&self) -> &SecretValue<Self::SecretLen>

Returns the private (secret) part of the keypair

fn public(&self) -> &Self::Public

Returns the public part of the keypair

Provided Methods

fn unzip(self) -> (SecretValue<Self::SecretLen>, Self::Public)

Consumes the instance and produces separated private and public parts