core_crypto::mls

Module conversation

MLS groups (aka conversation) are the actual entities cementing all the participants in a conversation.

This table summarizes what operations are permitted on a group depending its state: (PP=pending proposal, PC=pending commit)

can I ?	0 PP / 0 PC	1+ PP / 0 PC	0 PP / 1 PC	1+ PP / 1 PC

encrypt	✅	❌	❌	❌

handshake	✅	✅	❌	❌

merge	❌	❌	✅	✅

decrypt	✅	✅	✅	✅

Structs

BufferedDecryptedMessage

Type safe recursion of DecryptedMessage

CommitBundle

Returned when a commit is created

Conversation

A Conversation exposes the read-only interface of an MLS conversation.

ConversationConfiguration

The configuration parameters for a group/conversation

ConversationId

A unique identifier for a group/conversation. The identifier must be unique within a client.

ConversationIdRef

Reference to a ConversationId.

ConversationMut

A mutable view of an MLS conversation.

CustomConfiguration

The configuration parameters for a group/conversation which are not handled natively by openmls

DecryptedMessage

Represents the potential items a consumer might require after passing us an encrypted message we have decrypted for him

GroupInfoBundle

A [GroupInfo] with metadata

SecretKey

A secret key derived from the group secret.

WelcomeMessage

A Welcome Message as defined in RFC 9420.

Enums

Error

GroupInfoEncryptionType

GroupInfoEncryptionType

GroupInfoPayload

Represents the byte array in GroupInfoBundle

RatchetTreeType

RatchetTreeType

WirePolicy

Wrapper over [WireFormatPolicy]

Type Aliases

Result