core_crypto::mls::conversation

Module decrypt

MLS defines 3 kind of messages: Proposal, Commits and Application messages. Since they can (should) be all encrypted we need to first decrypt them before deciding what to do with them.

This table summarizes when a MLS group can decrypt any message:

can decrypt ?	0 pend. Commit	1 pend. Commit

0 pend. Proposal	✅	✅

1+ pend. Proposal	✅	✅

Structs

MlsBufferedConversationDecryptMessage

Type safe recursion of MlsConversationDecryptMessage

MlsConversationDecryptMessage

Represents the potential items a consumer might require after passing us an encrypted message we have decrypted for him