core_crypto::mls::conversation
Struct ConversationGuard

pub struct ConversationGuard { /* private fields */ }

A Conversation Guard wraps a GroupStoreValue<MlsConversation>.

By doing so, it permits mutable accesses to the conversation. This in turn means that we don’t have to duplicate the entire MlsConversation API on CentralContext.

Implementations

impl ConversationGuard

pub async fn add_members( &mut self, key_packages: Vec<KeyPackageIn>, ) -> Result<NewCrlDistributionPoint>

Adds new members to the group/conversation

pub async fn remove_members(&mut self, clients: &[ClientId]) -> Result<()>

Removes clients from the group/conversation.

Arguments

id - group/conversation id

clients - list of client ids to be removed from the group

pub async fn update_key_material(&mut self) -> Result<()>

Self updates the KeyPackage and automatically commits. Pending proposals will be commited.

Arguments

conversation_id - the group/conversation id

see [MlsCentral::update_keying_material]

pub async fn e2ei_rotate(&mut self, cb: Option<&CredentialBundle>) -> Result<()>

Send a commit in a conversation for changing the credential. Requires first having enrolled a new X509 certificate with either crate::context::CentralContext::e2ei_new_activation_enrollment or crate::context::CentralContext::e2ei_new_rotate_enrollment and having saved it with crate::context::CentralContext::save_x509_credential.

pub async fn commit_pending_proposals(&mut self) -> Result<()>

Commits all pending proposals of the group

impl ConversationGuard

pub async fn decrypt_message( &mut self, message: impl AsRef<[u8]>, ) -> Result<MlsConversationDecryptMessage>

Deserializes a TLS-serialized message, then processes it

Arguments

message - the encrypted message as a byte array

Returns

An MlsConversationDecryptMessage

Errors

If a message has been buffered, this will be indicated by an error. Other errors are originating from OpenMls and the KeyStore

impl ConversationGuard

pub async fn encrypt_message( &mut self, message: impl AsRef<[u8]>, ) -> Result<Vec<u8>>

Encrypts a raw payload then serializes it to the TLS wire format Can only be called when there is no pending commit and no pending proposal.

Arguments

message - the message as a byte array

Return type

This method will return an encrypted TLS serialized message.

Errors

If the conversation can’t be found, an error will be returned. Other errors are originating from OpenMls and the KeyStore

impl ConversationGuard

pub async fn clear_pending_proposal( &mut self, proposal_ref: MlsProposalRef, ) -> Result<()>

Allows to remove a pending (uncommitted) proposal. Use this when backend rejects the proposal you just sent e.g. if permissions have changed meanwhile.

CAUTION: only use this when you had an explicit response from the Delivery Service e.g. 403 or 409. Do not use otherwise e.g. 5xx responses, timeout etc..

Arguments

conversation_id - the group/conversation id

proposal_ref - unique proposal identifier which is present in crate::prelude::MlsProposalBundle and returned from all operation creating a proposal

Errors

When the conversation is not found or the proposal reference does not identify a proposal in the local pending proposal store

impl ConversationGuard

pub async fn wipe(&mut self) -> Result<()>

Destroys a group locally

Errors

KeyStore errors, such as IO

pub async fn mark_as_child_of( &mut self, parent_id: &ConversationId, ) -> Result<()>

Marks this conversation as child of another. Prerequisite: Must be a member of the parent group, and it must exist in the keystore