core_crypto::mls::conversation

Struct MlsConversation

pub struct MlsConversation { /* private fields */ }

This is a wrapper on top of the OpenMls’s MlsGroup, that provides Core Crypto specific functionality

This type will store the state of a group. With the MlsGroup it holds, it provides all operations that can be done in a group, such as creating proposals and commits. More information here

Implementations

impl MlsConversation

pub fn compute_next_commit_delay(&self) -> Option<u64>

Helps consumer by providing a deterministic delay in seconds for him to commit its pending proposal. It depends on the index of the client in the ratchet tree

self_index - ratchet tree index of self client

epoch - current group epoch

nb_members - number of clients in the group

impl MlsConversation

Creating proposals

pub async fn propose_add_member( &mut self, client: &Session, backend: &MlsCryptoProvider, key_package: KeyPackageIn, ) -> Result<MlsProposalBundle>

see openmls::group::MlsGroup::propose_add_member

pub async fn propose_remove_member( &mut self, client: &Session, backend: &MlsCryptoProvider, member: LeafNodeIndex, ) -> Result<MlsProposalBundle>

see openmls::group::MlsGroup::propose_remove_member

pub async fn propose_self_update( &mut self, client: &Session, backend: &MlsCryptoProvider, ) -> Result<MlsProposalBundle>

see openmls::group::MlsGroup::propose_self_update

pub async fn propose_explicit_self_update( &mut self, client: &Session, backend: &MlsCryptoProvider, leaf_node: Option<LeafNode>, ) -> Result<MlsProposalBundle>

see openmls::group::MlsGroup::propose_self_update

impl MlsConversation

pub async fn create( id: ConversationId, author_client: &Session, creator_credential_type: MlsCredentialType, configuration: MlsConversationConfiguration, backend: &MlsCryptoProvider, ) -> Result<Self>

Creates a new group/conversation

Arguments

id - group/conversation identifier

author_client - the client responsible for creating the group

creator_credential_type - kind of credential the creator wants to join the group with

config - group configuration

backend - MLS Provider that will be used to persist the group

Errors

Errors can happen from OpenMls or from the KeyStore

pub fn id(&self) -> &ConversationId

Group/conversation id

pub fn members(&self) -> HashMap<Vec<u8>, Credential>

Returns all members credentials from the group/conversation

pub fn members_in_next_epoch(&self) -> Vec<ClientId>

Get actual group members and subtract pending remove proposals

pub fn members_with_key(&self) -> HashMap<Vec<u8>, CredentialWithKey>

Returns all members credentials with their signature public key from the group/conversation