core_crypto::mls

Module credential

This module focuses on Credentials: cryptographic assertions of identity.

Credentials can be basic, or based on an x509 certificate chain.

Structs

Credential

A cryptographic credential.

CredentialRef

A reference to a credential which has been stored in a session.

FindFilters

Filters to narrow down the set of credentials returned from various credential-finding methods.

Enums

CredentialType

All supported Credential types.

Error