core_crypto::mls::credential

Enum Error

pub enum Error {
    DecodeX509(Error),
    InvalidIdentity,
    CredentialNotFound(SignaturePublicKey),
    UnsupportedCredentialType(u16),
    SignatureSchemeNotPresentInX509Identity(SignatureScheme),
    UnsupportedOperation(&'static str),
    UnsupportedAlgorithm,
    Keystore(KeystoreError),
    Mls(MlsError),
    Recursive(RecursiveError),
    TlsSerialize {
        source: Error,
        item: &'static str,
    },
    TlsDeserialize {
        source: Error,
        item: &'static str,
    },
}

Variants

DecodeX509(Error)

InvalidIdentity

CredentialNotFound(SignaturePublicKey)

UnsupportedCredentialType(u16)

Unsupported credential type.

Supported credential types:

basic

x509

SignatureSchemeNotPresentInX509Identity(SignatureScheme)

UnsupportedOperation(&'static str)

This operation is not supported.

There are some operations which must be implemented to satisfy a trait, but for which we cannot offer a real implementation. Those raise this error.

Where possible, a short workaround is included.

UnsupportedAlgorithm

Keystore(KeystoreError)

Mls(MlsError)

Recursive(RecursiveError)

TlsSerialize

Fields

source: Error

item: &'static str

TlsDeserialize

Fields

source: Error

item: &'static str

Implementations

impl Error

pub fn tls_serialize(item: &'static str) -> impl FnOnce(Error) -> Self

pub fn tls_deserialize(item: &'static str) -> impl FnOnce(Error) -> Self