core_crypto::mls::credential

Struct Credential

pub struct Credential { /* private fields */ }

A cryptographic credential.

This is tied to a particular client via either its client id or certificate bundle, depending on its credential type, but is independent of any client instance or storage.

To attach to a particular client instance and store, see TransactionContext::add_credential.

Implementations

impl Credential

pub fn x509( ciphersuite: Ciphersuite, cert: CertificateBundle, ) -> Result<Self, Error>

Create a new x509 credential from a certificate bundle.

impl Credential

pub fn basic( ciphersuite: Ciphersuite, client_id: ClientId, ) -> Result<Self, Error>

Generate a basic credential.

The result is independent of any client instance and the database; it lives in memory only.

The earliest validity of this credential is always 0. It will be updated once the credential is added to a session.

pub fn mls_credential(&self) -> &MlsCredential

Get the Openmls Credential type.

This stores the credential type (basic/x509).

pub fn credential_type(&self) -> CredentialType

Get the credential type

pub fn signature_scheme(&self) -> SignatureScheme

Get the signature scheme

pub fn ciphersuite(&self) -> Ciphersuite

Get the ciphersuite

pub fn to_mls_credential_with_key(&self) -> CredentialWithKey

Generate a CredentialWithKey, which combines the credential type with the public portion of the keypair.

pub fn earliest_validity(&self) -> u64

Earliest valid time of creation for this credential.

This is represented as seconds after the unix epoch.

Only meaningful for X509, where it is the โvalid_fromโ claim of the leaf credential. For basic credentials, this is always 0 when the credential is first created. It is updated upon being persisted to the database.

pub fn client_id(&self) -> &ClientIdRef

Get the client ID associated with this credential