Module core_crypto::prelude

Common imports that should be useful for most uses of the crate

Re-exports

pub use crate::e2e_identity::error::E2eIdentityError;

pub use crate::e2e_identity::error::E2eIdentityResult;

pub use crate::e2e_identity::types::E2eiAcmeChallenge;

pub use crate::e2e_identity::types::E2eiAcmeDirectory;

pub use crate::e2e_identity::types::E2eiNewAcmeAuthz;

pub use crate::e2e_identity::types::E2eiNewAcmeOrder;

pub use crate::e2e_identity::E2eiEnrollment;

pub use crate::mls::conversation::decrypt;

pub use crate::mls::conversation::decrypt::MlsBufferedConversationDecryptMessage;

pub use crate::mls::conversation::decrypt::MlsConversationDecryptMessage;

pub use crate::mls::conversation::ConversationId;

pub use crate::mls::conversation::MlsConversation;

pub use crate::mls::MlsCentral;

pub use crate::CoreCrypto;

pub use crate::CoreCryptoCallbacks;

Structs

CertificateBundle

Represents a x509 certificate chain supplied by the client It can fetch it after an end-to-end identity process where it can get back a certificate from the Authentication Service

Client

Represents a MLS client which in our case is the equivalent of a device. It can be the Android, iOS, web or desktop application which the authenticated user is using. A user has many client, a client has only one user. A client can belong to many MLS groups

ClientId

A unique identifier for clients. A client is an identifier for each App a user is using, such as desktop, mobile, etc. Users can have multiple clients. More information here

Credential

Credential.

EntropySeed

Wrapped 32-byte entropy seed with bounds check

GroupEpoch

Group epoch. Internally this is stored as a u64. The group epoch is incremented with every valid Commit that is merged into the group state.

KeyPackage

The key package struct.

KeyPackageIn

The key package struct.

MlsCentralConfiguration

Configuration parameters for MlsCentral

MlsCiphersuite

A wrapper for the OpenMLS Ciphersuite, so that we are able to provide a default value.

MlsCommitBundle

Returned when a commit is created

MlsConversationConfiguration

The configuration parameters for a group/conversation

MlsConversationCreationMessage

Returned when initializing a conversation through a commit. Different from conversation created from a [openmls::prelude::Welcome] message or an external commit.

MlsConversationInitBundle

Returned when a commit is created

MlsCryptoProvider

MlsCustomConfiguration

The configuration parameters for a group/conversation which are not handled natively by openmls

MlsGroup

A MlsGroup represents an MLS group with a high-level API. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

MlsGroupConfig

Specifies the configuration parameters for a MlsGroup. Refer to the User Manual for more information about the different configuration values.

MlsGroupInfoBundle

A [GroupInfo] with metadata

MlsMessageIn

Before use with the MlsGroup API, the message has to be unpacked via extract to yield its [MlsMessageInBody].

MlsProposalBundle

Returned when a Proposal is created. Helps roll backing a local proposal

MlsProposalRef

Abstraction over a [openmls::prelude::hash_ref::ProposalRef] to deal with conversions

MlsRotateBundle

Result returned after rotating the Credential of the current client in all the local conversations

VerifiableGroupInfo

A type that represents a group info of which the signature has not been verified. It implements the [Verifiable] trait and can be turned into a group info by calling verify(...) with the signature key of the Credential. When receiving a serialized group info, it can only be deserialized into a VerifiableGroupInfo, which can then be turned into a group info as described above.

WelcomeBundle

Contains everything client needs to know after decrypting an (encrypted) Welcome message

WireIdentity

Represents the identity claims identifying a client Those claims are verifiable by any member in the group

X509Identity

Represents the parts of WireIdentity that are specific to a X509 certificate (and not a Basic one). We don’t use an enum here since the sole purpose of this is to be exposed through the FFI (and union types are impossible to carry over the FFI boundary)

Enums

CiphersuiteName

MLS ciphersuites.

ClientIdentifier

Used by consumers to initializes a MLS client. Encompasses all the client types available. Could be enriched later with Verifiable Presentations.

CryptoError

CoreCrypto errors

CryptoboxMigrationError

Wrapper for errors that can happen during a Cryptobox migration

DeviceStatus

Indicates the standalone status of a device Credential in a MLS group at a moment T. This does not represent the states where a device is not using MLS or is not using end-to-end identity

E2eiConversationState

Indicates the state of a Conversation regarding end-to-end identity. Note: this does not check pending state (pending commit, pending proposals) so it does not consider members about to be added/removed

GroupInfoPayload

Represents the byte array in MlsGroupInfoBundle

MlsCredentialType

Lists all the supported Credential types. Could list in the future some types not supported by openmls such as Verifiable Presentation

MlsError

MLS-specific error wrapper - see github.com/openmls/openmls for details

MlsGroupInfoEncryptionType

GroupInfoEncryptionType

MlsProposal

Internal representation of proposal to ease further additions

MlsRatchetTreeType

RatchetTreeType

MlsWirePolicy

Wrapper over WireFormatPolicy

Node

Container enum for leaf and parent nodes.

ProteusError

Wrapper for Proteus-related errors

Constants

INITIAL_KEYING_MATERIAL_COUNT

Default number of KeyPackages a client generates the first time it’s created

Type Aliases

CryptoResult

A simpler definition for Result types that the Error is a CryptoError

KeyPackageRef

A reference to a key package. This value uniquely identifies a key package.

RawEntropySeed

32-byte raw entropy seed