core_crypto::prelude
Enum RecursiveError

pub enum RecursiveError {
    Root {
        context: &'static str,
        source: Box<Error>,
    },
    E2e {
        context: &'static str,
        source: Box<Error>,
    },
    Mls {
        context: &'static str,
        source: Box<Error>,
    },
    MlsClient {
        context: &'static str,
        source: Box<Error>,
    },
    MlsConversation {
        context: &'static str,
        source: Box<Error>,
    },
    MlsCredential {
        context: &'static str,
        source: Box<Error>,
    },
}

These errors wrap each of the module-specific errors in CoreCrypto.

The goal here is to reduce the need to redeclare each of these error types as an individual variant of a module-specific error type.

Variants

Root

Wrap a crate::Error for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

E2e

Wrap a crate::e2e_identity::Error for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

Mls

Wrap a crate::mls::Error for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

MlsClient

Wrap a [crate::mls::client::Error] for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

MlsConversation

Wrap a crate::mls::conversation::Error for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

MlsCredential

Wrap a [crate::mls::credential::Error] for recursion.

Fields

context: &'static str

What was happening in the caller

source: Box<Error>

What happened

Implementations

impl RecursiveError

pub fn root<E: Into<Error>>(context: &'static str) -> impl FnOnce(E) -> Self

Convert a crate::Error into a RecursiveError, with context

pub fn e2e_identity<E: Into<Error>>( context: &'static str, ) -> impl FnOnce(E) -> Self

Convert a crate::e2e_identity::Error into a RecursiveError, with context

pub fn mls<E: Into<Error>>(context: &'static str) -> impl FnOnce(E) -> Self

Convert a crate::mls::Error into a RecursiveError, with context

pub fn mls_client<E: Into<Error>>( context: &'static str, ) -> impl FnOnce(E) -> Self

Convert a [crate::mls::client::Error] into a RecursiveError, with context

pub fn mls_conversation<E: Into<Error>>( context: &'static str, ) -> impl FnOnce(E) -> Self

Convert a crate::mls::conversation::Error into a RecursiveError, with context

pub fn mls_credential<E: Into<Error>>( context: &'static str, ) -> impl FnOnce(E) -> Self

Convert a [crate::mls::credential::Error] into a RecursiveError, with context