Struct core_crypto::prelude::MlsConversation

pub struct MlsConversation { /* private fields */ }

This type will store the state of a group. With the MlsGroup it holds, it provides all operations that can be done in a group, such as creating proposals and commits. More information here

Implementations

impl MlsConversation

pub fn compute_next_commit_delay(&self) -> Option<u64>

Helps consumer by providing a deterministic delay in seconds for him to commit its pending proposal. It depends on the index of the client in the ratchet tree

self_index - ratchet tree index of self client

epoch - current group epoch

nb_members - number of clients in the group

impl MlsConversation

Abstraction over a MLS group capable of decrypting a MLS message

pub async fn decrypt_message(
&mut self,
message: impl AsRef<[u8]>,
backend: &MlsCryptoProvider,
callbacks: Option<&dyn CoreCryptoCallbacks>
) -> CryptoResult<MlsConversationDecryptMessage>

see MlsCentral::decrypt_message

impl MlsConversation

Abstraction over a MLS group capable of encrypting a MLS message

pub async fn encrypt_message(
&mut self,
message: impl AsRef<[u8]>,
backend: &MlsCryptoProvider
) -> CryptoResult<Vec<u8>>

see MlsCentral::encrypt_message It is durable because encrypting increments the message generation

impl MlsConversation

pub fn export_secret_key(
&self,
backend: &MlsCryptoProvider,
key_length: usize
) -> CryptoResult<Vec<u8>>

See MlsCentral::export_secret_key

pub fn get_client_ids(&self) -> Vec<ClientId>

See MlsCentral::get_client_ids

impl MlsConversation

Creating proposals

pub async fn propose_add_member(
&mut self,
backend: &MlsCryptoProvider,
key_package: &KeyPackage
) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_add_member

pub async fn propose_self_update(
&mut self,
backend: &MlsCryptoProvider
) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_self_update

pub async fn propose_remove_member(
&mut self,
backend: &MlsCryptoProvider,
member: &KeyPackageRef
) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_remove_member

impl MlsConversation

Abstraction over a MLS group capable of merging a commit

pub async fn commit_accepted(
&mut self,
backend: &MlsCryptoProvider
) -> CryptoResult<()>

see MlsCentral::commit_accepted

pub async fn clear_pending_proposal(
&mut self,
proposal_ref: MlsProposalRef,
backend: &MlsCryptoProvider
) -> CryptoResult<()>

see MlsCentral::clear_pending_proposal

pub async fn clear_pending_commit(
&mut self,
backend: &MlsCryptoProvider
) -> CryptoResult<()>

see MlsCentral::clear_pending_commit

impl MlsConversation

pub async fn create(
id: ConversationId,
author_client: &mut Client,
configuration: MlsConversationConfiguration,
backend: &MlsCryptoProvider
) -> CryptoResult<Self>

Creates a new group/conversation

Arguments

id - group/conversation identifier

author_client - the client responsible for creating the group

config - group configuration

backend - MLS Provider that will be used to persist the group

Errors

Errors can happen from OpenMls or from the KeyStore

pub async fn from_welcome_message(
welcome: Welcome,
configuration: MlsConversationConfiguration,
backend: &MlsCryptoProvider
) -> CryptoResult<Self>

Create the MLS conversation from an MLS Welcome message

Arguments

welcome - welcome message to create the group from

config - group configuration

backend - the KeyStore to persist the group

Errors

Errors can happen from OpenMls or from the KeyStore

pub fn id(&self) -> &ConversationId

Group/conversation id

pub fn members(&self) -> HashMap<MemberId, Vec<Credential>>

Returns all members credentials from the group/conversation

impl MlsConversation

pub fn members_in_next_epoch(
&self,
backend: &impl OpenMlsCrypto
) -> Vec<ClientId>

Get actual group members and subtract pending remove proposals