core_crypto::prelude
Struct MlsConversationDecryptMessage

pub struct MlsConversationDecryptMessage {
    pub app_msg: Option<Vec<u8>>,
    pub proposals: Vec<MlsProposalBundle>,
    pub is_active: bool,
    pub delay: Option<u64>,
    pub sender_client_id: Option<ClientId>,
    pub has_epoch_changed: bool,
    pub identity: WireIdentity,
    pub buffered_messages: Option<Vec<MlsBufferedConversationDecryptMessage>>,
    pub crl_new_distribution_points: NewCrlDistributionPoint,
}

Represents the potential items a consumer might require after passing us an encrypted message we have decrypted for him

Fields
app_msg: Option<Vec<u8>>

Decrypted text message

proposals: Vec<MlsProposalBundle>

Only when decrypted message is a commit, CoreCrypto will renew local proposal which could not make it in the commit. This will contain either:

local pending proposal not in the accepted commit

If there is a pending commit, its proposals which are not in the accepted commit

is_active: bool

Is the conversation still active after receiving this commit aka has the user been removed from the group

delay: Option<u64>

Delay time in seconds to feed caller timer for committing

sender_client_id: Option<ClientId>

ClientId of the sender of the message being decrypted. Only present for application messages.

has_epoch_changed: bool
๐Deprecated: This member will be removed in the future. Prefer using the EpochObserver interface.

Is the epoch changed after decrypting this message

identity: WireIdentity

Identity claims present in the sender credential Present for all messages

buffered_messages: Option<Vec<MlsBufferedConversationDecryptMessage>>

Only set when the decrypted message is a commit. Contains buffered messages for next epoch which were received before the commit creating the epoch because the DS did not fan them out in order.

crl_new_distribution_points: NewCrlDistributionPoint

New CRL distribution points that appeared by the introduction of a new credential