core_crypto::prelude

Struct MlsCryptoProvider

pub struct MlsCryptoProvider { /* private fields */ }

Implementations

impl MlsCryptoProvider

pub async fn try_new_with_configuration( config: MlsCryptoProviderConfiguration<'_>, ) -> Result<MlsCryptoProvider, MlsProviderError>

Initialize a CryptoProvider with a backend following the provided config (see: MlsCryptoProviderConfiguration)

pub async fn try_new( db_path: impl AsRef<str>, db_key: &DatabaseKey, ) -> Result<MlsCryptoProvider, MlsProviderError>

pub async fn try_new_in_memory( db_key: &DatabaseKey, ) -> Result<MlsCryptoProvider, MlsProviderError>

pub fn new_with_store( key_store: Connection, entropy_seed: Option<EntropySeed>, ) -> MlsCryptoProvider

Initialize a CryptoProvided with an already-configured backing store

pub async fn new_transaction(&self) -> Result<(), MlsProviderError>

Clones the references of the PkiEnvironment and the CryptoProvider into a transaction keystore to pass to openmls as the OpenMlsCryptoProvider

pub async fn update_pki_env( &self, pki_env: PkiEnvironment, ) -> Result<(), MlsProviderError>

Replaces the PKI env currently in place

pub async fn is_pki_env_setup(&self) -> bool

Returns whether we have a PKI env setup

pub fn reseed( &self, entropy_seed: Option<EntropySeed>, ) -> Result<(), MlsProviderError>

Reseeds the internal CSPRNG entropy pool with a brand new one.

If None is provided, the new entropy will be pulled through the current OS targetโs capabilities

pub async fn can_close(&self) -> bool

Returns whether or not it is currently possible to close this provider.

Reasons why it may not currently be possible:

A transaction is currently in progress

Multiple strong references currently exist to the keystore

As with all such checks, this is vulnerable to TOCTOU issues, but as the current implementation of the MlsCryptoProvider::close function consumes self, this is the only way to check in advance whether this will in principle work.

pub async fn close(self) -> Result<(), MlsProviderError>

Closes this provider, which in turns tears down the backing store

Note: This does not destroy the data on-disk in case of persistent backing store

pub fn keystore(&self) -> Connection

Clone keystore (its an Arc internnaly)

pub fn unwrap_keystore(self) -> Connection

Allows to retrieve the underlying key store directly