Trait core_crypto::primitives::DigestLike

pub trait DigestLike<T>where
    T: Update + FixedOutputReset + OutputSizeUser,{
    const SIZE: usize = <T::OutputSize>::USIZE;

    // Required method
    fn internal_state(&mut self) -> &mut T;

    // Provided methods
    fn update(&mut self, data: &[u8]) { ... }
    fn finalize_into(&mut self, out: &mut [u8]) { ... }
    fn finalize(&mut self) -> GenericArray<u8, T::OutputSize> { ... }
}

Generalization of digest-like operation (MAC, Digest,…) Defines the update and finalize operations to produce digest value of arbitrary data.

Provided Associated Constants

const SIZE: usize = <T::OutputSize>::USIZE

Length of the digest in bytes

Required Methods

fn internal_state(&mut self) -> &mut T

Access to the internal state of the digest-like operation.

Provided Methods

fn update(&mut self, data: &[u8])

Update the internal state of the digest-like using the given input data.

fn finalize_into(&mut self, out: &mut [u8])

Retrieve the final digest value into a prepared buffer and reset this instance so it could be reused for a new computation.

fn finalize(&mut self) -> GenericArray<u8, T::OutputSize>

Retrieve the final digest value and reset this instance so it could be reused for a new computation.