Struct core_crypto::proteus::ProteusCentral

pub struct ProteusCentral { /* private fields */ }

Proteus counterpart of crate::mls::MlsCentral The big difference is that ProteusCentral doesn’t own its own keystore but must borrow it from the outside. Whether it’s exclusively for this struct’s purposes or it’s shared with our main struct, [MlsCentral]

Implementations

impl ProteusCentral

pub async fn try_new(keystore: &CryptoKeystore) -> CryptoResult<Self>

Initializes the ProteusCentral

pub async fn session_from_prekey(
&mut self,
session_id: &str,
key: &[u8]
) -> CryptoResult<&mut ProteusConversationSession>

Creates a new session from a prekey

pub async fn session_from_message(
&mut self,
keystore: &mut CryptoKeystore,
session_id: &str,
envelope: &[u8]
) -> CryptoResult<(&mut ProteusConversationSession, Vec<u8>)>

Creates a new proteus Session from a received message

pub async fn session_save(
&self,
keystore: &CryptoKeystore,
session_id: &str
) -> CryptoResult<()>

Persists a session in store

Note: This isn’t usually needed as persisting sessions happens automatically when decrypting/encrypting messages and initializing Sessions

pub async fn session_delete(
&mut self,
keystore: &CryptoKeystore,
session_id: &str
) -> CryptoResult<()>

Deletes a session in the store

pub fn session_mut(
&mut self,
session_id: &str
) -> Option<&mut ProteusConversationSession>

Session mut accessor

pub fn session(&self, session_id: &str) -> Option<&ProteusConversationSession>

Session accessor

pub fn session_exists(&self, session_id: &str) -> bool

Session exists

pub async fn decrypt(
&mut self,
keystore: &mut CryptoKeystore,
session_id: &str,
ciphertext: &[u8]
) -> CryptoResult<Vec<u8>>

Decrypt a proteus message for an already existing session Note: This cannot be used for handshake messages, see ProteusCentral::session_from_message

pub async fn encrypt(
&mut self,
keystore: &mut CryptoKeystore,
session_id: &str,
plaintext: &[u8]
) -> CryptoResult<Vec<u8>>

Encrypt a message for a session

pub async fn encrypt_batched(
&mut self,
keystore: &mut CryptoKeystore,
sessions: &[impl AsRef<str>],
plaintext: &[u8]
) -> CryptoResult<HashMap<String, Vec<u8>>>

Encrypts a message for a list of sessions This is mainly used for conversations with multiple clients, this allows to minimize FFI roundtrips

pub async fn new_prekey(
&self,
id: u16,
keystore: &CryptoKeystore
) -> CryptoResult<Vec<u8>>

Generates a new Proteus PreKey, stores it in the keystore and returns a serialized PreKeyBundle to be consumed externally

pub async fn new_prekey_auto(
&self,
keystore: &CryptoKeystore
) -> CryptoResult<Vec<u8>>

Generates a new Proteus Prekey, with an automatically auto-incremented ID.

See ProteusCentral::new_prekey

pub fn identity(&self) -> &IdentityKeyPair

Proteus identity keypair

pub fn fingerprint(&self) -> String

Proteus Public key hex-encoded fingerprint

pub fn fingerprint_local(&self, session_id: &str) -> CryptoResult<String>

Proteus Session local hex-encoded fingerprint

Errors

When the session is not found

pub fn fingerprint_remote(&self, session_id: &str) -> CryptoResult<String>

Proteus Session remote hex-encoded fingerprint

Errors

When the session is not found

pub fn fingerprint_prekeybundle(prekey: &[u8]) -> CryptoResult<String>

Hex-encoded fingerprint of the given prekey

Errors

If the prekey cannot be deserialized

pub async fn cryptobox_migrate(
keystore: &CryptoKeystore,
path: &str
) -> CryptoResult<()>

Cryptobox -> CoreCrypto migration