Trait core_crypto::shared_keys::GroupElement

pub trait GroupElement<E: Scalar>: Clone + for<'a> Mul<&'a E, Output = Self> {
    type AlphaLen: ArrayLength<u8>;

    // Required methods
    fn to_alpha(&self) -> Alpha<Self::AlphaLen>;
    fn from_alpha(alpha: Alpha<Self::AlphaLen>) -> Result<Self>;
    fn generate(scalar: &E) -> Self;
    fn is_valid(&self) -> bool;

    // Provided methods
    fn random_pair() -> (Self, E) { ... }
    fn extract_key(&self, salt: &[u8]) -> SharedSecret { ... }
    fn expand_key(&self, salt: &[u8]) -> SharedSecret { ... }
}

Generic additive abelian group element with an associated scalar type. It also comes with the associated Alpha value size. A group element is considered valid if it is not neutral or a torsion element of small order.

Required Associated Types

type AlphaLen: ArrayLength<u8>

Length of the Alpha value - a binary representation of the group element.

Required Methods

fn to_alpha(&self) -> Alpha<Self::AlphaLen>

Converts the group element to a binary format suitable for representing the Alpha value.

fn from_alpha(alpha: Alpha<Self::AlphaLen>) -> Result<Self>

Converts the group element from the binary format representing an Alpha value.

fn generate(scalar: &E) -> Self

Create a group element using the group generator and the given scalar

fn is_valid(&self) -> bool

Group element is considered valid if it is not a neutral element and also not a torsion element of small order.

Provided Methods

fn random_pair() -> (Self, E)

Generates a random pair of group element and secret scalar. This is a convenience method that internally calls the random method of the associated Scalar and constructs the group element using generate.

fn extract_key(&self, salt: &[u8]) -> SharedSecret

Extract a keying material from a group element using HKDF extract

fn expand_key(&self, salt: &[u8]) -> SharedSecret

Performs KDF expansion from the given group element using HKDF expand