Trait core_crypto::shared_keys::SphinxSuite

pub trait SphinxSuite {
    type P: Keypair;
    type E: Scalar + for<'a> From<&'a Self::P>;
    type G: GroupElement<Self::E> + for<'a> From<&'a <Self::P as Keypair>::Public>;

    // Provided method
    fn new_shared_keys(
        public_keys: &[<Self::P as Keypair>::Public]
    ) -> Result<SharedKeys<Self::E, Self::G>> { ... }
}

Represents an instantiation of the Spinx protocol using the given EC group and corresponding public key object.

Required Associated Types

type P: Keypair

Keypair corresponding to the EC group

type E: Scalar + for<'a> From<&'a Self::P>

Scalar type supported by the EC group

type G: GroupElement<Self::E> + for<'a> From<&'a <Self::P as Keypair>::Public>

EC group element

Provided Methods

fn new_shared_keys( public_keys: &[<Self::P as Keypair>::Public] ) -> Result<SharedKeys<Self::E, Self::G>>

Convenience function to generate shared keys from the path of public keys.