core_crypto

Struct Database

pub struct Database { /* private fields */ }

Implementations

impl Database

pub async fn child_groups( &self, entity: PersistedMlsGroup, ) -> Result<Vec<PersistedMlsGroup>, CryptoKeystoreError>

pub async fn find_pending_messages_by_conversation_id( &self, conversation_id: &[u8], ) -> Result<Vec<MlsPendingMessage>, CryptoKeystoreError>

pub async fn remove_pending_messages_by_conversation_id( &self, conversation_id: impl AsRef<[u8]> + Send, ) -> Result<(), CryptoKeystoreError>

impl Database

These impls control the keystore transaction lifecycle.

pub async fn new_transaction(&self) -> Result<(), CryptoKeystoreError>

Waits for the current transaction to be committed or rolled back, then starts a new one.

pub async fn try_new_immediate_transaction( &self, ) -> Result<(), CryptoKeystoreError>

Start a new transaction if no other transaction is currently in progress.

If a transaction is currently in progress, this will produce a TransactionInProgress error.

pub async fn commit_transaction(&self) -> Result<(), CryptoKeystoreError>

pub async fn rollback_transaction(&self) -> Result<(), CryptoKeystoreError>

pub async fn transactionally<R>( &self, operation: impl AsyncFnOnce(&KeystoreTransaction) -> Result<R, CryptoKeystoreError>, ) -> Result<R, CryptoKeystoreError>

Do an operation on a keystore transaction on this database.

This is a convenience method abstracting over Self::new_transaction, Self::commit_transaction, and Self::rollback_transaction.

If the operation succeeds, the transaction is committed. Otherwise, it is rolled back.

impl Database

These impls are convenience methods to modify the keystore transaction.

Be aware that we are likely to remove these in the future: WPB-23951

pub async fn save<E>( &self, entity: E, ) -> Result<<E as EntityDatabaseMutation>::AutoGeneratedFields, CryptoKeystoreError>
where E: Entity + EntityDatabaseMutation + Send + Sync,

pub async fn remove<E>( &self, id: &<E as PrimaryKey>::PrimaryKey, ) -> Result<(), CryptoKeystoreError>
where E: Entity + EntityDatabaseMutation,

pub async fn remove_borrowed<E>( &self, id: &<E as BorrowPrimaryKey>::BorrowedPrimaryKey, ) -> Result<(), CryptoKeystoreError>
where E: Entity + EntityDatabaseMutation + BorrowPrimaryKey + EntityDeleteBorrowed,

impl Database

pub async fn open( path: &str, database_key: &DatabaseKey, ) -> Result<Arc<Database>, CryptoKeystoreError>

Open an encrypted Database at the provided location.

When compiled with target_os = "unknown", this opens a database encrypted via sqlite3-multiple-ciphers using its default encryption mechanism, stored in IndexedDB via the relaxed-idb shim.

When compiled normally, this opens a database encrypted via sqlcipher at a path in the local filesystem.

pub fn open_in_memory() -> Result<Arc<Database>, CryptoKeystoreError>

Open an in-memory Database.

In-memory databases are never encrypted.

pub async fn update_key( &self, new_key: &DatabaseKey, ) -> Result<(), CryptoKeystoreError>

Change the encryption key for this database.

pub async fn close(self) -> Result<(), CryptoKeystoreError>

pub async fn wipe(self) -> Result<(), CryptoKeystoreError>

Close and remove this database.

This deletes the database, including its encryption key. Future opens will always succeed with any arbitrary encryption key; they will simply open an empty database.

pub async fn location(&self) -> Option<String>

Get the location of the database.

Returns None if the database is in-memory.

pub async fn export_copy( &self, destination_path: &str, ) -> Result<(), CryptoKeystoreError>

Export a copy of the database to the specified path using VACUUM INTO.

This creates a fully vacuumed and optimized copy of the database. The copy will be encrypted with the same key as the source database.

Arguments

destination_path - The file path where the database copy should be created

impl Database

pub async fn mls_fetch_key_packages<V>( &self, count: u32, ) -> Result<Vec<V>, CryptoKeystoreError>
where V: MlsEntity,

Fetches Keypackages

Arguments

count - amount of entries to be returned

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_group_exists(&self, group_id: impl AsRef<[u8]> + Send) -> bool

Checks if the given MLS group id exists in the keystore Note: in case of any error, this will return false

Arguments

group_id - group/conversation id

pub async fn mls_group_persist( &self, group_id: impl AsRef<[u8]> + Send, state: &[u8], parent_group_id: Option<&[u8]>, ) -> Result<(), CryptoKeystoreError>

Persists a MlsGroup

Arguments

group_id - group/conversation id

state - the group state

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_groups_restore( &self, ) -> Result<HashMap<Vec<u8>, (Option<Vec<u8>>, Vec<u8>)>, CryptoKeystoreError>

Loads MlsGroups from the database. It will be returned as a HashMap where the key is the group/conversation id and the value the group state

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_group_delete( &self, group_id: impl AsRef<[u8]> + Send, ) -> Result<(), CryptoKeystoreError>

Deletes MlsGroups from the database.

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_pending_groups_save( &self, group_id: impl AsRef<[u8]> + Send, mls_group: &[u8], custom_configuration: &[u8], parent_group_id: Option<&[u8]>, ) -> Result<(), CryptoKeystoreError>

Saves a MlsGroup in a temporary table (typically used in scenarios where the group cannot be committed until the backend acknowledges it, like external commits)

Arguments

group_id - group/conversation id

mls_group - the group/conversation state

custom_configuration - local group configuration

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_pending_groups_load( &self, group_id: impl AsRef<[u8]> + Send, ) -> Result<Option<(Vec<u8>, Vec<u8>)>, CryptoKeystoreError>

Loads a temporary MlsGroup and its configuration from the database

Arguments

id - group/conversation id

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn mls_pending_groups_delete( &self, group_id: impl AsRef<[u8]> + Send, ) -> Result<(), CryptoKeystoreError>

Deletes a temporary MlsGroup from the database

Arguments

id - group/conversation id

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

pub async fn save_e2ei_enrollment( &self, id: &[u8], content: &[u8], ) -> Result<(), CryptoKeystoreError>

Persists an enrollment instance

Arguments

id - hash of the enrollment and unique identifier

content - serialized enrollment

pub async fn pop_e2ei_enrollment( &self, id: &[u8], ) -> Result<Option<Vec<u8>>, CryptoKeystoreError>

Fetches and delete the enrollment instance

Arguments

id - hash of the enrollment and unique identifier

impl Database

pub async fn proteus_store_prekey( &self, id: u16, prekey: &[u8], ) -> Result<(), CryptoKeystoreError>