core_crypto

Trait InnermostErrorMessage

pub trait InnermostErrorMessage {
    // Required method
    fn innermost_error_message(&self) -> String;
}

Produce the error message from the innermost wrapped error.

We produce arbitrarily nested errors which are very helpful at capturing relevant context, and very bad at surfacing the root error cause in a default .to_string() call.

This trait, automatically implemented for all standard errors, rectifies this problem.

Required Methods

fn innermost_error_message(&self) -> String

Produce the error message from the innermost wrapped error.