core_crypto

Module transaction_context

This module contains the primitives to enable transactional support on a higher level within the Session. All mutating operations need to be done through a TransactionContext.

Modules

conversation

This module contains all super::TransactionContext methods related to a conversation.

e2e_identity

This module contains all behaviour of a transaction context connected to end-to-end identity.

key_package

This module contains all transactional behavior related to key packages

proteus

This module contains all super::TransactionContext methods concerning proteus.

Structs

TransactionContext

This struct provides transactional support for Core Crypto.

Enums

Error

Errors produced during a transaction

Type Aliases

Result

A module-specific Result type with a default error variant.