core_crypto::transaction_context

Struct TransactionContext

pub struct TransactionContext { /* private fields */ }

This struct provides transactional support for Core Crypto.

This struct provides mutable access to the internals of Core Crypto. Every operation that causes data to be persisted needs to be done through this struct. This struct will buffer all operations in memory and when TransactionContext::finish is called, it will persist the data into the keystore.

Implementations

impl TransactionContext

pub async fn join_by_external_commit( &self, group_info: VerifiableGroupInfo, credential_ref: &CredentialRef, ) -> Result<WelcomeBundle>

Issues an external commit and stores the group in a temporary table. This method is intended for example when a new client wants to join the user’s existing groups. On success this function will return the group id and a message to be fanned out to other clients.

If the Delivery Service accepts the external commit, you have to ensure the commit is merged in order to get back a functional MLS group. If it rejects it, you can retry by calling Self::join_by_external_commit again.

Arguments

group_info - a GroupInfo wrapped in a MLS message. it can be obtained by deserializing a TLS serialized GroupInfo object

credential_ref - reference to the crate::Credential to use for joining this group.

Returns WelcomeBundle

Errors

Errors resulting from OpenMls, the KeyStore calls and serialization

impl TransactionContext

pub async fn new_external_add_proposal( &self, conversation_id: ConversationId, epoch: GroupEpoch, ciphersuite: Ciphersuite, credential_type: CredentialType, ) -> Result<MlsMessageOut>

Crafts a new external Add proposal. Enables a client outside a group to request addition to this group. For Wire only, the client must belong to an user already in the group

Arguments

conversation_id - the group/conversation id

epoch - the current epoch of the group. See openmls::group::GroupEpoch

ciphersuite - of the new openmls::prelude::KeyPackage to create

credential_type - of the new openmls::prelude::KeyPackage to create

Return type

Returns a message with the proposal to be add a new client

Errors

Errors resulting from the creation of the proposal within OpenMls. Fails when credential_type is CredentialType::X509 and no Credential has been created for it beforehand with TransactionContext::e2ei_mls_init_only or variants.

impl TransactionContext

pub async fn new_add_proposal( &self, id: &ConversationId, key_package: KeyPackage, ) -> Result<MlsProposalBundle>

Creates a new Add proposal

pub async fn new_remove_proposal( &self, id: &ConversationId, client_id: ClientId, ) -> Result<MlsProposalBundle>

Creates a new Add proposal

pub async fn new_update_proposal( &self, id: &ConversationId, ) -> Result<MlsProposalBundle>

Creates a new Add proposal

impl TransactionContext

pub async fn process_raw_welcome_message( &self, welcome: &[u8], ) -> Result<WelcomeBundle>

Create a conversation from a TLS serialized MLS Welcome message. The MlsConversationConfiguration used in this function will be the default implementation.

Arguments

welcome - a TLS serialized welcome message

configuration - configuration of the MLS conversation fetched from the Delivery Service

Return type

This function will return the conversation/group id

Errors

see TransactionContext::process_welcome_message

pub async fn process_welcome_message( &self, welcome: MlsMessageIn, ) -> Result<WelcomeBundle>

Create a conversation from a received MLS Welcome message

Arguments

welcome - a Welcome message received as a result of a commit adding new members to a group

configuration - configuration of the group/conversation

Return type

This function will return the conversation/group id

Errors

Errors can be originating from the KeyStore of from OpenMls:

if no openmls::key_packages::KeyPackage can be read from the KeyStore

if the message can’t be decrypted

impl TransactionContext

pub async fn conversation( &self, id: &ConversationIdRef, ) -> Result<ConversationGuard>

Acquire a conversation guard.

This helper struct permits mutations on a conversation.

pub async fn new_conversation( &self, id: &ConversationIdRef, credential_ref: &CredentialRef, config: MlsConversationConfiguration, ) -> Result<()>

Create a new empty conversation

Arguments

id - identifier of the group/conversation (must be unique otherwise the existing group will be overridden)

creator_credential_type - kind of credential the creator wants to create the group with

config - configuration of the group/conversation

Errors

Errors can happen from the KeyStore or from OpenMls for ex if no openmls::key_packages::KeyPackage can be found in the KeyStore

pub async fn conversation_exists(&self, id: &ConversationIdRef) -> Result<bool>

Checks if a given conversation id exists locally

impl TransactionContext

pub async fn e2ei_verify_group_state( &self, group_info: VerifiableGroupInfo, ) -> Result<E2eiConversationState>

See crate::mls::session::Session::e2ei_verify_group_state.

pub async fn get_credential_in_use( &self, group_info: VerifiableGroupInfo, credential_type: CredentialType, ) -> Result<E2eiConversationState>

See crate::mls::session::Session::get_credential_in_use.

impl TransactionContext

pub async fn e2ei_is_enabled( &self, signature_scheme: SignatureScheme, ) -> Result<bool>

See crate::mls::session::Session::e2ei_is_enabled

impl TransactionContext

pub async fn e2ei_is_pki_env_setup(&self) -> Result<bool>

See crate::mls::session::Session::e2ei_is_pki_env_setup.

pub async fn e2ei_register_acme_ca( &self, trust_anchor_pem: String, ) -> Result<()>

Registers a Root Trust Anchor CA for the use in E2EI processing.

Please note that without a Root Trust Anchor, all validations will fail; So this is the first step to perform after initializing your E2EI client

Parameters

trust_anchor_pem - PEM certificate to anchor as a Trust Root

pub async fn e2ei_register_intermediate_ca_pem( &self, cert_pem: String, ) -> Result<NewCrlDistributionPoints>

Registers an Intermediate CA for the use in E2EI processing.

Please note that a Root Trust Anchor CA is needed to validate Intermediate CAs; You need to have a Root CA registered before calling this

Parameters

cert_pem - PEM certificate to register as an Intermediate CA

pub async fn e2ei_register_crl( &self, crl_dp: String, crl_der: Vec<u8>, ) -> Result<CrlRegistration>

Registers a CRL for the use in E2EI processing.

Please note that a Root Trust Anchor CA is needed to validate CRLs; You need to have a Root CA registered before calling this

Parameters

crl_dp - CRL Distribution Point; Basically the URL you fetched it from

crl_der - DER representation of the CRL

Returns

A CrlRegistration with the dirty state of the new CRL (see struct) and its expiration timestamp

impl TransactionContext

pub async fn e2ei_new_activation_enrollment( &self, display_name: String, handle: String, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment>

Generates an E2EI enrollment instance for a “regular” client (with a Basic credential) willing to migrate to E2EI. As a consequence, this method does not support changing the ClientId which should remain the same as the Basic one. Once the enrollment is finished, use the instance in TransactionContext::save_x509_credential to save the new credential.

pub async fn e2ei_new_rotate_enrollment( &self, display_name: Option<String>, handle: Option<String>, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment>

Generates an E2EI enrollment instance for a E2EI client (with a X509 certificate credential) having to change/rotate their credential, either because the former one is expired or it has been revoked. As a consequence, this method does not support changing neither ClientId which should remain the same as the previous one. It lets you change the DisplayName or the handle if you need to. Once the enrollment is finished, use the instance in TransactionContext::save_x509_credential to do the rotation.

pub async fn save_x509_credential( &self, enrollment: &mut E2eiEnrollment, certificate_chain: String, ) -> Result<NewCrlDistributionPoints>

Saves a new X509 credential. Requires first having enrolled a new X509 certificate with either TransactionContext::e2ei_new_activation_enrollment or TransactionContext::e2ei_new_rotate_enrollment.

Expected actions to perform after this function (in this order)

Rotate credentials for each conversation in crate::mls::conversation::ConversationGuard::e2ei_rotate

Generate new key packages with Self::generate_keypackage

Use these to replace the stale ones the in the backend

Delete the old credentials and keypackages locally using Self::remove_credential

impl TransactionContext

pub async fn e2ei_enrollment_stash( &self, enrollment: E2eiEnrollment, ) -> Result<Vec<u8>>

Allows persisting an active enrollment (for example while redirecting the user during OAuth) in order to resume it later with TransactionContext::e2ei_enrollment_stash_pop

Arguments

enrollment - the enrollment instance to persist

Returns

A handle for retrieving the enrollment later on

pub async fn e2ei_enrollment_stash_pop( &self, handle: Vec<u8>, ) -> Result<E2eiEnrollment>

Fetches the persisted enrollment and deletes it from the keystore

Arguments

handle - returned by TransactionContext::e2ei_enrollment_stash

impl TransactionContext

pub async fn e2ei_new_enrollment( &self, client_id: ClientId, display_name: String, handle: String, team: Option<String>, expiry_sec: u32, ciphersuite: Ciphersuite, ) -> Result<E2eiEnrollment>

Creates an enrollment instance with private key material you can use in order to fetch a new x509 certificate from the acme server.

Parameters

client_id - client identifier e.g. b7ac11a4-8f01-4527-af88-1c30885a7931:6add501bacd1d90e@example.com

display_name - human readable name displayed in the application e.g. Smith, Alice M (QA)

handle - user handle e.g. alice.smith.qa@example.com

expiry_sec - generated x509 certificate expiry in seconds

pub async fn e2ei_mls_init_only( &self, enrollment: &mut E2eiEnrollment, certificate_chain: String, transport: Arc<dyn MlsTransport>, ) -> Result<NewCrlDistributionPoints>

Parses the ACME server response from the endpoint fetching x509 certificates and uses it to initialize the MLS client with a certificate

impl TransactionContext

pub async fn generate_keypackage( &self, credential_ref: &CredentialRef, lifetime: Option<Duration>, ) -> Result<Keypackage>

Generate a Keypackage from the referenced credential.

Makes no attempt to look up or prune existing keypackges.

If lifetime is set, the keypackages will expire that span into the future. If it is unset, KEYPACKAGE_DEFAULT_LIFETIME is used.

pub async fn get_keypackage_refs(&self) -> Result<Vec<KeypackageRef>>

Get all KeypackageRefs known to the keystore.

pub async fn remove_keypackage(&self, kp_ref: &KeypackageRef) -> Result<()>

Remove a Keypackage from the keystore.

pub async fn remove_keypackages_for( &self, credential_ref: &CredentialRef, ) -> Result<()>

Remove all Keypackages associated with this ref.

impl TransactionContext

pub async fn proteus_init(&self) -> Result<()>

Initializes the proteus client

pub async fn proteus_reload_sessions(&self) -> Result<()>

Reloads the sessions from the key store

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or it will do nothing

pub async fn proteus_session_from_prekey( &self, session_id: &str, prekey: &[u8], ) -> Result<Arc<RwLock<ProteusConversationSession>>>

Creates a proteus session from a prekey

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session_from_message( &self, session_id: &str, envelope: &[u8], ) -> Result<(Arc<RwLock<ProteusConversationSession>>, Vec<u8>)>

Creates a proteus session from a Proteus message envelope

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session_save(&self, session_id: &str) -> Result<()>

Saves a proteus session in the keystore

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session_delete(&self, session_id: &str) -> Result<()>

Deletes a proteus session from the keystore

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session( &self, session_id: &str, ) -> Result<Option<Arc<RwLock<ProteusConversationSession>>>>

Proteus session accessor

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session_exists(&self, session_id: &str) -> Result<bool>

Proteus session exists

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_decrypt( &self, session_id: &str, ciphertext: &[u8], ) -> Result<Vec<u8>>

Decrypts a proteus message envelope

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_encrypt( &self, session_id: &str, plaintext: &[u8], ) -> Result<Vec<u8>>

Encrypts proteus message for a given session ID

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_encrypt_batched( &self, sessions: &[impl AsRef<str>], plaintext: &[u8], ) -> Result<HashMap<String, Vec<u8>>>

Encrypts a proteus message for several sessions ID. This is more efficient than other methods as the calls are batched. This also reduces the rountrips when crossing over the FFI

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_new_prekey(&self, prekey_id: u16) -> Result<Vec<u8>>

Creates a new Proteus prekey and returns the CBOR-serialized version of the prekey bundle

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_new_prekey_auto(&self) -> Result<(u16, Vec<u8>)>

Creates a new Proteus prekey with an automatically incremented ID and returns the CBOR-serialized version of the prekey bundle

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_last_resort_prekey(&self) -> Result<Vec<u8>>

Returns the last resort prekey

pub fn proteus_last_resort_prekey_id() -> u16

Returns the proteus last resort prekey id (u16::MAX = 65535)

pub async fn proteus_fingerprint(&self) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_fingerprint_local( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_fingerprint_remote( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with TransactionContext::proteus_init first or an error will be returned

impl TransactionContext

pub async fn mls_provider(&self) -> Result<MlsCryptoProvider>

Clones all references that the MlsCryptoProvider comprises.

pub async fn finish(&self) -> Result<()>

Commits the transaction, meaning it takes all the enqueued operations and persist them into the keystore. After that the internal state is switched to invalid, causing errors if something is called from this object.

pub async fn abort(&self) -> Result<()>

Aborts the transaction, meaning it discards all the enqueued operations. After that the internal state is switched to invalid, causing errors if something is called from this object.

pub async fn mls_init( &self, identifier: ClientIdentifier, ciphersuites: &[Ciphersuite], transport: Arc<dyn MlsTransport>, ) -> Result<()>

Initializes the MLS client of super::CoreCrypto.

pub async fn client_public_key( &self, ciphersuite: Ciphersuite, credential_type: CredentialType, ) -> Result<Vec<u8>>

Returns the client’s public key.

pub async fn client_id(&self) -> Result<ClientId>

see Session::id

pub async fn random_bytes(&self, len: usize) -> Result<Vec<u8>>

Generates a random byte array of the specified size

pub async fn set_data(&self, data: Vec<u8>) -> Result<()>

Set arbitrary data to be retrieved by TransactionContext::get_data. This is meant to be used as a check point at the end of a transaction. The data should be limited to a reasonable size.

pub async fn get_data(&self) -> Result<Option<Vec<u8>>>

Get the data that has previously been set by TransactionContext::set_data. This is meant to be used as a check point at the end of a transaction.

pub async fn add_credential( &self, credential: Credential, ) -> Result<CredentialRef>

Add a credential to the identities of this session.

As a side effect, stores the credential in the keystore.

pub async fn remove_credential( &self, credential_ref: &CredentialRef, ) -> Result<()>

Remove a credential from the identities of this session.

As a side effect, delete the credential from the keystore.

Removes both the credential itself and also any key packages which were generated from it.

pub async fn find_credentials( &self, find_filters: CredentialFindFilters<'_>, ) -> Result<Vec<CredentialRef>>

Find credentials matching the find filters among the identities of this session

Note that finding credentials with no filters set is equivalent to Self::get_credentials.

pub async fn get_credentials(&self) -> Result<Vec<CredentialRef>>

Get all credentials from the identities of this session.

To get specific credentials, it can be more efficient to use Self::find_credentials.