Struct core_crypto::types::PublicKey

pub struct PublicKey { /* private fields */ }

Represents a secp256k1 public key. This is a “Schrödinger public key”, both compressed and uncompressed to save some cycles.

Implementations

impl PublicKey

pub fn random() -> Self

Generates a new random public key. Because the corresponding private key is discarded, this might be useful only for testing purposes.

pub fn to_address(&self) -> Address

Converts the public key to an Ethereum address

pub fn to_bytes(&self, compressed: bool) -> Box<[u8]>

Serializes the public key to a binary form.

pub fn to_hex(&self, compressed: bool) -> String

Serializes the public key to a binary form and converts it to hexadecimal string representation.

impl PublicKey

pub const SIZE_COMPRESSED: usize = 33usize

Size of the compressed public key in bytes

pub const SIZE_UNCOMPRESSED_PLAIN: usize = 64usize

pub const SIZE_UNCOMPRESSED: usize = 65usize

Size of the uncompressed public key in bytes

pub fn random_keypair() -> ([u8; 32], PublicKey)

Generates new random keypair (private key, public key)

pub fn from_bytes(data: &[u8]) -> Result<Self>

pub fn from_privkey(private_key: &[u8]) -> Result<PublicKey>

pub fn from_signature(msg: &[u8], signature: &Signature) -> Result<PublicKey>

pub fn from_signature_hash( hash: &[u8], signature: &Signature ) -> Result<PublicKey>

pub fn combine(summands: &[&PublicKey]) -> PublicKey

Sums all given public keys together, creating a new public key. Panics if reaches infinity (EC identity point), which is an invalid public key.

pub fn tweak_add(key: &PublicKey, tweak: &[u8]) -> PublicKey

Adds the given public key with tweak times secp256k1 generator, producing a new public key. Panics if reaches infinity (EC identity point), which is an invalid public key.