Struct core_crypto::types::VrfParameters

pub struct VrfParameters {
    pub v: CurvePoint,
    pub h: Scalar,
    pub s: Scalar,
    pub h_v: CurvePoint,
    pub s_b: CurvePoint,
}

Bundles values given to the smart contract to prove that a ticket is a win.

The VRF is thereby needed because it generates on-demand determinitstic entropy that can only be derived by the ticket redeemer.

Fields
v: CurvePoint

the pseudo-random point

h: Scalars: Scalarh_v: CurvePoint

helper value for smart contract

s_b: CurvePoint

helper value for smart contract

Implementations

impl VrfParameters

pub fn verify<const T: usize>( &self, creator: &Address, msg: &[u8; T], dst: &[u8] ) -> Result<()>

Verifies that VRF values are valid