core_crypto::mls

Module conversation

MLS groups (aka conversation) are the actual entities cementing all the participants in a conversation.

This table summarizes what operations are permitted on a group depending its state: (PP=pending proposal, PC=pending commit)

can I ?	0 PP / 0 PC	1+ PP / 0 PC	0 PP / 1 PC	1+ PP / 1 PC

encrypt	✅	❌	❌	❌

handshake	✅	✅	❌	❌

merge	❌	❌	✅	✅

decrypt	✅	✅	✅	✅

Structs

ConversationGuard

A Conversation Guard wraps a GroupStoreValue<MlsConversation>.

ImmutableConversation

An ImmutableConversation wraps a MlsConversation.

MlsConversation

This is a wrapper on top of the OpenMls’s MlsGroup, that provides Core Crypto specific functionality

Enums

Error

Traits

Conversation

The Conversation trait provides a set of operations that can be done on an immutable conversation.

Type Aliases

ConversationId

A unique identifier for a group/conversation. The identifier must be unique within a client.

Result