core_crypto::mls::conversation

Enum DecryptedMessage

pub enum DecryptedMessage {
    Text(Text),
    Commit(Commit),
    Proposal(Proposal),
}

Represents the potential items a consumer might require after passing us an encrypted message we have decrypted for him

Variants

Text(Text)

The decrypted message is a text message.

Commit(Commit)

The decrypted message is a commit.

Proposal(Proposal)

The decrypted message is a proposal.

Implementations

impl DecryptedMessage

pub fn is_text(&self) -> bool

Returns true if this is a DecryptedMessage::Text, otherwise false

pub fn as_text_mut(&mut self) -> Option<&mut Text>

Optionally returns mutable references to the inner fields if this is a DecryptedMessage::Text, otherwise None

pub fn as_text(&self) -> Option<&Text>

Optionally returns references to the inner fields if this is a DecryptedMessage::Text, otherwise None

pub fn into_text(self) -> Result<Text, Self>

Returns the inner fields if this is a DecryptedMessage::Text, otherwise returns back the enum in the Err case of the result

pub unsafe fn into_text_unchecked(self) -> Text

Unchecked return of the inner fields of DecryptedMessage::Text.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_text_unchecked(&self) -> &Text

Unchecked reference of the inner fields of DecryptedMessage::Text.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_text_mut_unchecked(&mut self) -> &mut Text

Unchecked mutable reference of the inner fields of DecryptedMessage::Text.

Safety

Results in undefined behavior when it is the incorrect variant.

pub fn is_commit(&self) -> bool

Returns true if this is a DecryptedMessage::Commit, otherwise false

pub fn as_commit_mut(&mut self) -> Option<&mut Commit>

Optionally returns mutable references to the inner fields if this is a DecryptedMessage::Commit, otherwise None

pub fn as_commit(&self) -> Option<&Commit>

Optionally returns references to the inner fields if this is a DecryptedMessage::Commit, otherwise None

pub fn into_commit(self) -> Result<Commit, Self>

Returns the inner fields if this is a DecryptedMessage::Commit, otherwise returns back the enum in the Err case of the result

pub unsafe fn into_commit_unchecked(self) -> Commit

Unchecked return of the inner fields of DecryptedMessage::Commit.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_commit_unchecked(&self) -> &Commit

Unchecked reference of the inner fields of DecryptedMessage::Commit.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_commit_mut_unchecked(&mut self) -> &mut Commit

Unchecked mutable reference of the inner fields of DecryptedMessage::Commit.

Safety

Results in undefined behavior when it is the incorrect variant.

pub fn is_proposal(&self) -> bool

Returns true if this is a DecryptedMessage::Proposal, otherwise false

pub fn as_proposal_mut(&mut self) -> Option<&mut Proposal>

Optionally returns mutable references to the inner fields if this is a DecryptedMessage::Proposal, otherwise None

pub fn as_proposal(&self) -> Option<&Proposal>

Optionally returns references to the inner fields if this is a DecryptedMessage::Proposal, otherwise None

pub fn into_proposal(self) -> Result<Proposal, Self>

Returns the inner fields if this is a DecryptedMessage::Proposal, otherwise returns back the enum in the Err case of the result

pub unsafe fn into_proposal_unchecked(self) -> Proposal

Unchecked return of the inner fields of DecryptedMessage::Proposal.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_proposal_unchecked(&self) -> &Proposal

Unchecked reference of the inner fields of DecryptedMessage::Proposal.

Safety

Results in undefined behavior when it is the incorrect variant.

pub unsafe fn as_proposal_mut_unchecked(&mut self) -> &mut Proposal

Unchecked mutable reference of the inner fields of DecryptedMessage::Proposal.

Safety

Results in undefined behavior when it is the incorrect variant.

impl DecryptedMessage

pub fn identity(&self) -> &WireIdentity

Identity claims present in the sender credential