core_crypto::mls::conversation

Struct MlsConversation

pub struct MlsConversation { /* private fields */ }

This is a wrapper on top of the OpenMls’s MlsGroup, that provides Core Crypto specific functionality

This type will store the state of a group. With the MlsGroup it holds, it provides all operations that can be done in a group, such as creating proposals and commits. More information here

Implementations

impl MlsConversation

pub fn compute_next_commit_delay(&self) -> Option<u64>

Helps consumer by providing a deterministic delay in seconds for him to commit its pending proposal. It depends on the index of the client in the ratchet tree

self_index - ratchet tree index of self client

epoch - current group epoch

nb_members - number of clients in the group

impl MlsConversation

Abstraction over a MLS group capable of decrypting a MLS message

pub async fn decrypt_message( &mut self, message: MlsMessageIn, parent_conv: Option<&Arc<RwLock<MlsConversation>>>, client: &Client, backend: &MlsCryptoProvider, callbacks: Option<&dyn CoreCryptoCallbacks>, restore_pending: bool, ) -> CryptoResult<MlsConversationDecryptMessage>

see CentralContext::decrypt_message

impl MlsConversation

Abstraction over a MLS group capable of encrypting a MLS message

pub async fn encrypt_message( &mut self, client: &Client, message: impl AsRef<[u8]>, backend: &MlsCryptoProvider, ) -> CryptoResult<Vec<u8>>

see CentralContext::encrypt_message It is durable because encrypting increments the message generation

impl MlsConversation

pub fn export_secret_key( &self, backend: &impl OpenMlsCryptoProvider, key_length: usize, ) -> CryptoResult<Vec<u8>>

See MlsCentral::export_secret_key

pub fn get_client_ids(&self) -> Vec<ClientId>

See MlsCentral::get_client_ids

impl MlsConversation

Abstraction over a MLS group capable of merging a commit

pub async fn commit_accepted( &mut self, backend: &MlsCryptoProvider, ) -> CryptoResult<()>

see CentralContext::commit_accepted

pub async fn clear_pending_proposal( &mut self, proposal_ref: MlsProposalRef, backend: &MlsCryptoProvider, ) -> CryptoResult<()>

see CentralContext::clear_pending_proposal

pub async fn clear_pending_commit( &mut self, backend: &MlsCryptoProvider, ) -> CryptoResult<()>

see CentralContext::clear_pending_commit

impl MlsConversation

Creating proposals

pub async fn propose_add_member( &mut self, client: &Client, backend: &MlsCryptoProvider, key_package: KeyPackageIn, ) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_add_member

pub async fn propose_remove_member( &mut self, client: &Client, backend: &MlsCryptoProvider, member: LeafNodeIndex, ) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_remove_member

pub async fn propose_self_update( &mut self, client: &Client, backend: &MlsCryptoProvider, ) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_self_update

pub async fn propose_explicit_self_update( &mut self, client: &Client, backend: &MlsCryptoProvider, leaf_node: Option<LeafNode>, ) -> CryptoResult<MlsProposalBundle>

see openmls::group::MlsGroup::propose_self_update

impl MlsConversation

pub async fn create( id: ConversationId, author_client: &Client, creator_credential_type: MlsCredentialType, configuration: MlsConversationConfiguration, backend: &MlsCryptoProvider, ) -> CryptoResult<Self>

Creates a new group/conversation

Arguments

id - group/conversation identifier

author_client - the client responsible for creating the group

creator_credential_type - kind of credential the creator wants to join the group with

config - group configuration

backend - MLS Provider that will be used to persist the group

Errors

Errors can happen from OpenMls or from the KeyStore

pub fn id(&self) -> &ConversationId

Group/conversation id

pub fn members(&self) -> HashMap<Vec<u8>, Credential>

Returns all members credentials from the group/conversation

pub fn members_with_key(&self) -> HashMap<Vec<u8>, CredentialWithKey>

Returns all members credentials with their signature public key from the group/conversation

pub async fn mark_as_child_of( &mut self, parent_id: &ConversationId, keystore: &Connection, ) -> CryptoResult<()>

Marks this conversation as child of another. Prequisite: Being a member of this group and for it to be stored in the keystore

impl MlsConversation

pub fn members_in_next_epoch(&self) -> Vec<ClientId>

Get actual group members and subtract pending remove proposals