Struct core_crypto::mls::MlsCentral

pub struct MlsCentral { /* private fields */ }

The entry point for the MLS CoreCrypto library. This struct provides all functionality to create and manage groups, make proposals and commits.

Implementations

impl MlsCentral

pub async fn decrypt_message(
&mut self,
conversation_id: &ConversationId,
message: impl AsRef<[u8]>
) -> CryptoResult<MlsConversationDecryptMessage>

Deserializes a TLS-serialized message, then deciphers it

Arguments

conversation - the group/conversation id

message - the encrypted message as a byte array

Return type

This method will return a tuple containing an optional message and an optional delay time for the callers to wait for committing. A message will be None in case the provided payload in case of a system message, such as Proposals and Commits. Otherwise it will return the message as a byte array. The delay will be Some when the message has a proposal

Errors

If the conversation can’t be found, an error will be returned. Other errors are originating from OpenMls and the KeyStore

impl MlsCentral

pub async fn encrypt_message(
&mut self,
conversation: &ConversationId,
message: impl AsRef<[u8]>
) -> CryptoResult<Vec<u8>>

Encrypts a raw payload then serializes it to the TLS wire format

Arguments

conversation - the group/conversation id

message - the message as a byte array

Return type

This method will return an encrypted TLS serialized message.

Errors

If the conversation can’t be found, an error will be returned. Other errors are originating from OpenMls and the KeyStore

impl MlsCentral

pub fn export_secret_key(
&self,
conversation_id: &ConversationId,
key_length: usize
) -> CryptoResult<Vec<u8>>

Derives a new key from the one in the group, allowing it to be use elsewehere.

Arguments

conversation_id - the group/conversation id

key_length - the length of the key to be derived. If the value is higher than the bounds of u16 or the context hash * 255, an error will be returned

Errors

OpenMls secret generation error or conversation not found

pub fn get_client_ids(
&self,
conversation_id: &ConversationId
) -> CryptoResult<Vec<ClientId>>

Exports the clients from a conversation

Arguments

conversation_id - the group/conversation id

Errors

if the conversation can’t be found

impl MlsCentral

pub async fn add_members_to_conversation(
&mut self,
id: &ConversationId,
members: &mut [ConversationMember]
) -> CryptoResult<MlsConversationCreationMessage>

Adds new members to the group/conversation

Arguments

id - group/conversation id

members - members to be added to the group

Return type

An optional struct containing a welcome and a message will be returned on successful call. The value will be None only if the group can’t be found locally (no error will be returned in this case).

Errors

If the authorisation callback is set, an error can be caused when the authorization fails. Other errors are KeyStore and OpenMls errors:

pub async fn remove_members_from_conversation(
&mut self,
id: &ConversationId,
clients: &[ClientId]
) -> CryptoResult<MlsCommitBundle>

Removes clients from the group/conversation.

Arguments

id - group/conversation id

clients - list of client ids to be removed from the group

Return type

An optional message will be returned on successful call. The value will be None only if the group can’t be found locally (no error will be returned in this case).

Errors

If the authorisation callback is set, an error can be caused when the authorization fails. Other errors are KeyStore and OpenMls errors.

pub async fn update_keying_material(
&mut self,
id: &ConversationId
) -> CryptoResult<MlsCommitBundle>

Self updates the KeyPackage and automatically commits. Pending proposals will be commited

Arguments

conversation_id - the group/conversation id

Return type

A tuple containing the message with the commit this call generated and an optional welcome message that will be present if there were pending add proposals to be commited

Errors

If the conversation can’t be found, an error will be returned. Other errors are originating from OpenMls and the KeyStore

pub async fn commit_pending_proposals(
&mut self,
id: &ConversationId
) -> CryptoResult<Option<MlsCommitBundle>>

Commits all pending proposals of the group

Arguments

backend - the KeyStore to persist group changes

Return type

A tuple containing the commit message and a possible welcome (in the case Add proposals were pending within the internal MLS Group)

Errors

Errors can be originating from the KeyStore and OpenMls

impl MlsCentral

A MLS group is a distributed object scattered across many parties. We use a Delivery Service to orchestrate those parties. So when we create a commit, a mutable operation, it has to be validated by the Delivery Service. But it might occur that another group member did the exact same thing at the same time. So if we arrive second in this race, we must “rollback” the commit we created and accept (“merge”) the other one. A client would

Create a commit

Send the commit to the Delivery Service

When Delivery Service responds

200 OK –> use MlsCentral::commit_accepted to merge the commit

409 CONFLICT –> do nothing. MlsCentral::decrypt_message will restore the proposals not committed

5xx –> retry

pub async fn commit_accepted(
&mut self,
conversation_id: &ConversationId
) -> CryptoResult<()>

The commit we created has been accepted by the Delivery Service. Hence it is guaranteed to be used for the new epoch. We can now safely “merge” it (effectively apply the commit to the group) and update it in the keystore. The previous can be discarded to respect Forward Secrecy.

pub async fn clear_pending_proposal(
&mut self,
conversation_id: &ConversationId,
proposal_ref: Vec<u8>
) -> CryptoResult<()>

Allows to remove a pending (uncommitted) proposal. Use this when backend rejects the proposal you just sent e.g. if permissions have changed meanwhile.

CAUTION: only use this when you had an explicit response from the Delivery Service e.g. 403 or 409. Do not use otherwise e.g. 5xx responses, timeout etc..

Arguments

conversation_id - the group/conversation id

proposal_ref - unique proposal identifier which is present in [MlsProposalBundle] and returned from all operation creating a proposal

Errors

When the conversation is not found or the proposal reference does not identify a proposal in the local pending proposal store

pub async fn clear_pending_commit(
&mut self,
conversation_id: &ConversationId
) -> CryptoResult<()>

Allows to remove a pending commit. Use this when backend rejects the commit you just sent e.g. if permissions have changed meanwhile.

CAUTION: only use this when you had an explicit response from the Delivery Service e.g. 403. Do not use otherwise e.g. 5xx responses, timeout etc.. DO NOT use when Delivery Service responds 409, pending state will be renewed in MlsCentral::decrypt_message

Arguments

conversation_id - the group/conversation id

Errors

When the conversation is not found or there is no pending commit

impl MlsCentral

pub async fn wipe_conversation(
&mut self,
conversation_id: &ConversationId
) -> CryptoResult<()>

Destroys a group locally

Errors

KeyStore errors, such as IO

impl MlsCentral

pub async fn join_by_external_commit(
&self,
public_group_state: VerifiablePublicGroupState,
custom_cfg: MlsCustomConfiguration
) -> CryptoResult<MlsConversationInitBundle>

Issues an external commit and stores the group in a temporary table. This method is intended for example when a new client wants to join the user’s existing groups. On success this function will return the group id and a message to be fanned out to other clients.

If the Delivery Service accepts the external commit, you have to [merge_pending_group_from_external_commit] in order to get back a functional MLS group. On the opposite, if it rejects it, you can either retry by just calling again [join_by_external_commit], no need to [clear_pending_group_from_external_commit]. If you want to abort the operation (too many retries or the user decided to abort), you can use [clear_pending_group_from_external_commit] in order not to bloat the user’s storage but nothing bad can happen if you forget to except some storage space wasted.

Arguments

group_state - a verifiable public group state. it can be obtained by deserializing a TLS serialized PublicGroupState object

custom_cfg - configuration of the MLS conversation fetched from the Delivery Service

Return type

It will return a tuple with the group/conversation id and the message containing the commit that was generated by this call

Errors

Errors resulting from OpenMls, the KeyStore calls and serialization

pub async fn merge_pending_group_from_external_commit(
&mut self,
id: &ConversationId
) -> CryptoResult<()>

This merges the commit generated by [join_by_external_commit], persists the group permanently and deletes the temporary one. After merging, the group should be fully functional.

Arguments

id - the conversation id

Errors

Errors resulting from OpenMls, the KeyStore calls and deserialization

pub async fn clear_pending_group_from_external_commit(
&mut self,
id: &ConversationId
) -> CryptoResult<()>

In case the external commit generated by [join_by_external_commit] is rejected by the Delivery Service and we want to abort this external commit once for all, we can wipe out the pending group from the keystore in order not to waste space

Arguments

id - the conversation id

Errors

Errors resulting from the KeyStore calls

impl MlsCentral

pub async fn new_external_add_proposal(
&self,
conversation_id: ConversationId,
epoch: GroupEpoch
) -> CryptoResult<MlsMessageOut>

Crafts a new external Add proposal. Enables a client outside a group to request addition to this group. For Wire only, the client must belong to an user already in the group

Arguments

conversation_id - the group/conversation id

epoch - the current epoch of the group. See openmls::group::GroupEpoch

Return type

Returns a message with the proposal to be add a new client

Errors

Errors resulting from the creation of the proposal within OpenMls

pub async fn new_external_remove_proposal(
&self,
conversation_id: ConversationId,
epoch: GroupEpoch,
key_package_ref: KeyPackageRef
) -> CryptoResult<MlsMessageOut>

Crafts a new external Remove proposal. Enables a client outside a group to request removal of a client within the group.

Arguments

conversation_id - the group/conversation id

epoch - the current epoch of the group. See openmls::group::GroupEpoch

key_package_ref - the KeyPackageRef of the client to be added to the group

Return type

Returns a message with the proposal to be remove a client

Errors

Errors resulting from the creation of the proposal within OpenMls

impl MlsCentral

pub async fn new_proposal(
&mut self,
conversation: &ConversationId,
proposal: MlsProposal
) -> CryptoResult<MlsProposalBundle>

Creates a new proposal within a group

Arguments

conversation - the group/conversation id

proposal - the proposal do be added in the group

Return type

A [ProposalBundle] with the proposal in a Mls message and a reference to that proposal in order to rollback it if required

Errors

If the conversation is not found, an error will be returned. Errors from OpenMls can be returned as well, when for example there’s a commit pending to be merged

impl MlsCentral

pub async fn try_new(
configuration: MlsCentralConfiguration,
certificate_bundle: Option<CertificateBundle>
) -> CryptoResult<Self>

Tries to initialize the MLS Central object. Takes a store path (i.e. Disk location of the embedded database, should be consistent between messaging sessions) And a root identity key (i.e. enclaved encryption key for this device)

Arguments

configuration - the configuration for the MlsCentral

certificate_bundle - an optional CertificateBundle. It will be used to generate the CredentialBundle. If Noneis passed, credentials of type a Basic will be created, otherwise a x509.

Errors

Failures in the initialization of the KeyStore can cause errors, such as IO, the same kind of errors can happen when the groups are being restored from the KeyStore or even during the client initialization (to fetch the identity signature). Other than that, MlsError can be caused by group deserialization or during the initialization of the credentials:

for x509 Credentials if the cetificate chain length is lower than 2

for Basic Credentials if the signature key cannot be generated either by not supported scheme or the key generation fails

pub async fn try_new_in_memory(
configuration: MlsCentralConfiguration,
certificate_bundle: Option<CertificateBundle>
) -> CryptoResult<Self>

Same as the [crate::MlsCentral::try_new] but instead, it uses an in memory KeyStore. Although required, the store_path parameter from the MlsCentralConfiguration won’t be used here.

pub async fn mls_init(
&mut self,
client_id: ClientId,
ciphersuites: Vec<MlsCiphersuite>,
certificate_bundle: Option<CertificateBundle>
) -> CryptoResult<()>

Initializes the MLS client if [CoreCrypto] has previously been initialized with [CoreCrypto::deferred_init] instead of [CoreCrypto::new]. This should stay as long as proteus is supported. Then it should be removed.

pub async fn restore_from_disk(&mut self) -> CryptoResult<()>

MlsCentral is supposed to be a singleton. Knowing that, it does some optimizations by keeping MLS groups in memory. Sometimes, especially on iOS, it is required to use extensions to perform tasks in the background. Extensions are executed in another process so another MlsCentral instance has to be used. This method has to be used to synchronize instances. It simply fetches the MLS group from keystore in memory.

pub fn callbacks(&mut self, callbacks: Box<dyn CoreCryptoCallbacks>)

Sets the consumer callbacks (i.e authorization callbacks for CoreCrypto to perform authorization calls when needed)

Arguments

callbacks - a callback to be called to perform authorization

pub fn client_public_key(&self) -> CryptoResult<Vec<u8>>

Returns the client’s public key as a buffer

pub fn client_id(&self) -> CryptoResult<ClientId>

Returns the client’s id as a buffer

pub async fn client_keypackages(
&self,
amount_requested: usize
) -> CryptoResult<Vec<KeyPackageBundle>>

Returns amount_requested OpenMLS [KeyPackageBundle]s. Will always return the requested amount as it will generate the necessary (lacking) amount on-the-fly

Note: Keypackage pruning is performed as a first step

Arguments

amount_requested - number of KeyPackages to request and fill the KeyPackageBundle

Return type

A vector of KeyPackageBundle

Errors

Errors can happen when accessing the KeyStore

pub async fn client_valid_keypackages_count(&self) -> CryptoResult<usize>

Returns the count of valid, non-expired, unclaimed keypackages in store

pub async fn new_conversation(
&mut self,
id: ConversationId,
config: MlsConversationConfiguration
) -> CryptoResult<()>

Create a new empty conversation

Arguments

id - identifier of the group/conversation (must be unique otherwise the existing group will be overridden)

config - configuration of the group/conversation

Errors

Errors can happen from the KeyStore or from OpenMls for ex if no [KeyPackageBundle] can be found in the KeyStore

pub fn conversation_exists(&self, id: &ConversationId) -> bool

Checks if a given conversation id exists locally

pub fn conversation_epoch(&self, id: &ConversationId) -> CryptoResult<u64>

Returns the epoch of a given conversation

Errors

If the conversation can’t be found

pub async fn process_welcome_message(
&mut self,
welcome: Welcome,
custom_cfg: MlsCustomConfiguration
) -> CryptoResult<ConversationId>

Create a conversation from a received MLS Welcome message

Arguments

welcome - a Welcome message received as a result of a commit adding new members to a group

configuration - configuration of the group/conversation

Return type

This function will return the conversation/group id

Errors

Errors can be originating from the KeyStore of from OpenMls:

if no [KeyPackageBundle] can be read from the KeyStore

if the message can’t be decrypted

pub async fn process_raw_welcome_message(
&mut self,
welcome: Vec<u8>,
custom_cfg: MlsCustomConfiguration
) -> CryptoResult<ConversationId>

Create a conversation from a TLS serialized MLS Welcome message. The MlsConversationConfiguration used in this function will be the default implementation.

Arguments

welcome - a TLS serialized welcome message

configuration - configuration of the MLS conversation fetched from the Delivery Service

Return type

This function will return the conversation/group id

Errors

see MlsCentral::process_welcome_message

pub async fn export_public_group_state(
&self,
conversation_id: &ConversationId
) -> CryptoResult<Vec<u8>>

Exports a TLS-serialized view of the current group state corresponding to the provided conversation ID.

Arguments

conversation - the group/conversation id

message - the encrypted message as a byte array

Return type

A TLS serialized byte array of the PublicGroupState

Errors

If the conversation can’t be found, an error will be returned. Other errors are originating from OpenMls and serialization

pub async fn close(self) -> CryptoResult<()>

Closes the connection with the local KeyStore

Errors

KeyStore errors, such as IO

pub async fn wipe(self) -> CryptoResult<()>

Destroys everything we have, in-memory and on disk.

Errors

KeyStore errors, such as IO

pub fn random_bytes(&self, len: usize) -> CryptoResult<Vec<u8>>

Generates a random byte array of the specified size

pub fn provider(&self) -> &MlsCryptoProvider

Returns a reference for the internal Crypto Provider

pub fn provider_mut(&mut self) -> &mut MlsCryptoProvider

Returns a mutable reference for the internal Crypto Provider