core_crypto

Struct CoreCrypto

pub struct CoreCrypto { /* private fields */ }

Wrapper superstruct for both mls::session::Session and proteus::ProteusCentral

As std::ops::Deref is implemented, this struct is automatically dereferred to mls::session::Session apart from proteus_* calls

This is cheap to clone as all internal members have Arc wrappers or are Copy.

Implementations

impl CoreCrypto

pub async fn history_client(history_secret: HistorySecret) -> Result<Self>

Instantiate a history client.

This client exposes the full interface of CoreCrypto, but it should only be used to decrypt messages. Other use is a logic error.

impl CoreCrypto

pub async fn proteus_session( &self, session_id: &str, ) -> Result<Option<Arc<RwLock<ProteusConversationSession>>>>

Proteus session accessor

Warning: The Proteus client MUST be initialized with crate::transaction_context::TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_session_exists(&self, session_id: &str) -> Result<bool>

Proteus session exists

Warning: The Proteus client MUST be initialized with crate::transaction_context::TransactionContext::proteus_init first or an error will be returned

pub fn proteus_last_resort_prekey_id() -> u16

Returns the proteus last resort prekey id (u16::MAX = 65535)

pub async fn proteus_fingerprint(&self) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with crate::transaction_context::TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_fingerprint_local( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with crate::transaction_context::TransactionContext::proteus_init first or an error will be returned

pub async fn proteus_fingerprint_remote( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with crate::transaction_context::TransactionContext::proteus_init first or an error will be returned

impl CoreCrypto

pub async fn new_transaction(&self) -> Result<TransactionContext>

Creates a new transaction. All operations that persist data will be buffered in memory and when TransactionContext::finish is called, the data will be persisted in a single database transaction.

impl CoreCrypto

pub fn new(database: Database) -> Self

Create an new CoreCrypto client without any initialized session.

pub async fn set_pki_environment( &self, pki_environment: Option<PkiEnvironment>, ) -> Result<()>

Set the session’s PKI Environment

pub async fn get_pki_environment(&self) -> Option<PkiEnvironment>

Get the session’s PKI Environment

pub async fn mls_session(&self) -> Result<Session<Database>>

Get the mls session if initialized

pub fn database(&self) -> Database

Get the database

pub async fn close(&self) -> Result<()>

Closes the database indexdb connections must be closed explicitly while rusqlite implements drop which suffices.