core_crypto

Struct CoreCrypto

pub struct CoreCrypto { /* private fields */ }

Wrapper superstruct for both [mls::Client] and proteus::ProteusCentral

As std::ops::Deref is implemented, this struct is automatically dereferred to [mls::Client] apart from proteus_* calls

This is cheap to clone as all internal members have Arc wrappers or are Copy.

Implementations

impl CoreCrypto

pub async fn register_epoch_observer( &self, epoch_observer: Arc<dyn EpochObserver>, ) -> Result<(), Error>

Add an epoch observer to this session.

This function should be called 0 or 1 times in a session’s lifetime. If called when an epoch observer already exists, this will return an error.

impl CoreCrypto

pub async fn proteus_session( &self, session_id: &str, ) -> Result<Option<Arc<RwLock<ProteusConversationSession>>>>

Proteus session accessor

Warning: The Proteus client MUST be initialized with [TransactionContext::proteus_init] first or an error will be returned

pub async fn proteus_session_exists(&self, session_id: &str) -> Result<bool>

Proteus session exists

Warning: The Proteus client MUST be initialized with [TransactionContext::proteus_init] first or an error will be returned

pub fn proteus_last_resort_prekey_id() -> u16

Returns the proteus last resort prekey id (u16::MAX = 65535)

pub async fn proteus_fingerprint(&self) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with [TransactionContext::proteus_init] first or an error will be returned

pub async fn proteus_fingerprint_local( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with [TransactionContext::proteus_init] first or an error will be returned

pub async fn proteus_fingerprint_remote( &self, session_id: &str, ) -> Result<String>

Returns the proteus identity’s public key fingerprint

Warning: The Proteus client MUST be initialized with [TransactionContext::proteus_init] first or an error will be returned

impl CoreCrypto

pub async fn new_transaction(&self) -> Result<TransactionContext>

Creates a new transaction. All operations that persist data will be buffered in memory and when TransactionContext::finish is called, the data will be persisted in a single database transaction.

impl CoreCrypto

pub fn take(self) -> Session

Allows to extract the MLS Client from the wrapper superstruct