core_crypto
Trait MlsTransport

pub trait MlsTransport:
    Debug
    + Send
    + Sync {
    // Required methods
    fn send_commit_bundle<'life0, 'async_trait>(
        &'life0 self,
        commit_bundle: MlsCommitBundle,
    ) -> Pin<Box<dyn Future<Output = Result<MlsTransportResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_message<'life0, 'async_trait>(
        &'life0 self,
        mls_message: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<MlsTransportResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Client callbacks to allow communication with the delivery service. There are two different endpoints, one for messages and one for commit bundles.

Required Methods

fn send_commit_bundle<'life0, 'async_trait>( &'life0 self, commit_bundle: MlsCommitBundle, ) -> Pin<Box<dyn Future<Output = Result<MlsTransportResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send a commit bundle to the corresponding endpoint.

fn send_message<'life0, 'async_trait>( &'life0 self, mls_message: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<MlsTransportResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send a message to the corresponding endpoint.