core_crypto

Trait ToRecursiveError

pub trait ToRecursiveError {
    // Required method
    fn construct_recursive(self, context: &'static str) -> RecursiveError;
}

Like Into, but different, because we don’t actually want to implement Into for our subordinate error types.

By forcing ourselves to map errors everywhere in order for question mark operators to work, we ensure that

Required Methods

fn construct_recursive(self, context: &'static str) -> RecursiveError

Construct a recursive error given the current context